The BLOSUM62_enc matrix

The BLOck SUbstition Matrix (BLOSUM) for sequences with less than 62% similarity. The matrix has been subset to the 20 proteogenic amino acids.

Format

A data frame with with 21 rows and 21 columns.

and also row and column zero vectors representing 'X' (any amino acid). You can check three-letter codes and full names of the amino acids in aminoacids_df.

Source

Henikoff, Steven, and Jorja G. Henikoff. "Amino acid substitution matrices from protein blocks." Proceedings of the National Academy of Sciences 89.22 (1992): 10915-10919. ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM62

Details

BLOSUM matrix is a substitution matrix used for sequence alignment of proteins. BLOSUM matrices are actual percentage identity values of sequences selected for construction of the matrices. BLOSUM62 indicates that the sequences selected for constructing the matrix share an average identity value of 62%. BLOSUM62 is miderange matrix between close and distantly related proteins. Matrix made by matblas from blosum62.iij BLOSUM Clustered Scoring Matrix in 1/2 Bit Units Cluster Percentage: >= 62 Entropy = 0.6979, Expected = -0.5209

See also

Examples

data(BLOSUM62_enc)