The BLOSUM62_pca matrix

This matrix enables principal component analysis (PCA) with usage BLOSUM62.

Format

A data frame with with 21 rows and 20 columns.

and ordinal numbers of principal component direction as columns names. Twenty columns represent the loadings of the twenty eigenvectors. You can check three-letter codes and full names of the amino acids in aminoacids_df.

Source

Li, Jie & Koehl, Patrice. (2014). 3D representations of amino acids - Applications to protein sequence comparison and classification. Computational and Structural Biotechnology Journal. 11. 10.1016/j.csbj.2014.09.001. ftp://ftp.ncbi.nih.gov/blast/matrices/BLOSUM62

Details

The BLOSUM62_pca matrix enables PCA calculation on proteins sequences alignments. Components are generated by an eigenvector decomposition of the matrix formed from pairwise similarity scores between each pair of sequences. The similarity score model used for creating BLOSUM62_pca matrix is the BLOSUM62.

See also

Examples

data(BLOSUM62_pca)