Amino acid abbreviations

A dataset containing the one- and three-letter codes, and full names of amino acids according to the IUPAC nomenclature and HGVS Recommendations for the Description of Sequence Variants.

Format

A data frame with 27 rows and 4 columns:

one

One-letter codes of amino acids

three

Three-letter codes of amino acids

full

Full name of the amino acid

amb

Logical indicating if abbreviation is ambiguous, i. e., matches more than one amino acid

Source

(1984), Nomenclature and Symbolism for Amino Acids and Peptides. European Journal of Biochemistry 138, 9-37; Dunnen et al. (2016), HGVS Recommendations for the Description of Sequence Variants: 2016 Update. Human Mutation 37, 564-569.

Details

The dataset contains an amino acid alphabet of one-letter codes with the corresponding three-letter codes and full names of the amino acids. It also includes a gap symbol '-' that can be found in sequence alignments and termination symbol '*' used to indicate the end of translation.

Examples

data(aminoacids_df)