Concatenate sq objects

Merges multiple sq objects of the same type into one larger.

# S3 method for sq
c(...)

Arguments

...
multiple sq objects. All of them have to have the same type and subtype. If type is atp, unt or enc also their alphabets have to be exactly identical.

Value

A sq object with length equal to sum of lengths of individual objects passed as parameters. Elements of sq are concatenated just as if they were normal lists (see c)

Examples

sq_1 <- construct_sq(c("TAGACTAG", "", "CCGTAGATG"))
sq_2 <- construct_sq(c("TTGATAACG", "TGTATGTGA"))
sq_3 <- construct_sq(character(0))
sq_4 <- construct_sq("gaGG")

c(sq_1, sq_2, sq_3, sq_4)

#> dna (DNA), cln (cleaned) sequences list:
#> [1] TAGACTAG                                                                 <8>
#> [2] <NULL>                                                                   <0>
#> [3] CCGTAGATG                                                                <9>
#> [4] TTGATAACG                                                                <9>
#> [5] TGTATGTGA                                                                <9>
#> [6] GAGG                                                                     <4>