Compare sq object

Compares input sq object with another given.

# S3 method for sq
==(x1, x2)

Arguments

x1
a sq object.

x2
an object (character vector or sq object) to compare with sq.

Value

A logical vector indicating on which positions objects are the same

Details

`==` converts left hand side of comparison (x1) to characters vector using as.character and checks whether given on the right side object can be compared with sq object. Function also check the type of sq object with which given object will be compared. If the type of sq object is ami or nuc and given sequence is character vector consisting lowercase, the function rewrites it into capital ones with usage toupper. If right hand side object (x2) is sq it is converted to character vector using also as.character function.

When both objects are already converted to character vectors, comparison is done elementwise with standard R rules, (e.g. recycling is used). You can check details here.

Comparing sequences as characters vectors cause that various types of sequences can be compared for example amino acids with nucleotides sequences so attention should be paid which sequences types are compared.

See also

Examples

# Creating sq object to work on:
sq <- construct_sq(c("ACTGCTG", "CTTAGA",
                     "CCCT", "CTGAATGT"), type = "nuc")

sq_different_len <- construct_sq(c("ACTGCTG", "CTTAGA",
                                   "GGAA"), type = "nuc")

sq_the_same_len <- construct_sq(c("ACTGCTG", "CTTAGA",
                                 "CCCT", "CTGAATGT"), type = "nuc")

# Get an overview of the sequences:
summary(sq)

#>      Length Class  Mode
#> [1,] 3      -none- raw 
#> [2,] 3      -none- raw 
#> [3,] 2      -none- raw 
#> [4,] 3      -none- raw 

summary(sq_the_same_len)

#>      Length Class  Mode
#> [1,] 3      -none- raw 
#> [2,] 3      -none- raw 
#> [3,] 2      -none- raw 
#> [4,] 3      -none- raw 

summary(sq_different_len)

#>      Length Class  Mode
#> [1,] 3      -none- raw 
#> [2,] 3      -none- raw 
#> [3,] 2      -none- raw 

# Comparing sq object with an object of the same length :
sq == sq_the_same_len

#> [1] TRUE TRUE TRUE TRUE

# Comparing object sq object with an object of a different length : 
sq == sq_different_len

#> Warning: longer object length is not a multiple of shorter object length

#> [1]  TRUE  TRUE FALSE FALSE

# Comparing sq object to given character vector of a different length:
sq == c('AAA','CCC')

#> [1] FALSE FALSE FALSE FALSE

# Comparing sq object to given character vector of a the same length:
sq == c("ACTGCTG", "CTTAGA",'CCCT', 'CTGAATGT')

#> [1] TRUE TRUE TRUE TRUE

# Comparing sq object to given nucleotide element 'ATGTGA':
sq == 'ATGTGA'

#> [1] FALSE FALSE FALSE FALSE

# Comparing sq object to given amino acids vector:
sq == c('RISGQQD','RISGQQD')

#> [1] FALSE FALSE FALSE FALSE