Export sq objects into other formats

Convert object of class sq to another class from another package. Currently supported packages are ape with its formats (AAbin and DNAbin), Bioconductor (AAStringSet, DNAStringSet) and seqinr (SeqFastaAA, SeqFastadna).

export_sq(sq, export_format, name)

Arguments

sq
a sq object.

export_format
a character string indicating package and the destination class; it should be one of the following: "seqinr::SeqFastaAA", "ape::AAbin", "Biostrings::AAStringSet", "seqinr::SeqFastadna", "ape::DNAbin", "Biostrings::DNAStringSet".

name
a character vector of length equal to sq length.

See also

Examples

sq_ami <- construct_sq(c("MVVGL", "LAVPP"))
export_sq(sq_ami, "ape::AAbin")

#> 2 amino acid sequences in a list
#> 
#> All sequences of the same length: 5 
#> 

export_sq(sq_ami, "Biostrings::AAStringSet", c("one", "two"))

#>   A AAStringSet instance of length 2
#>     width seq                                               names               
#> [1]     5 MVVGL                                             one
#> [2]     5 LAVPP                                             two

export_sq(sq_ami, "seqinr::SeqFastaAA")

#> [[1]]
#> [1] "M" "V" "V" "G" "L"
#> attr(,"class")
#> [1] "SeqFastaAA"
#> 
#> [[2]]
#> [1] "L" "A" "V" "P" "P"
#> attr(,"class")
#> [1] "SeqFastaAA"
#> 

sq_nuc <- construct_sq(c("TGATGAAGCGCA", "TTGATGGGAA"))
export_sq(sq_nuc, "ape::DNAbin", name = c("one", "two"))

#> 2 DNA sequences in binary format stored in a list.
#> 
#> Mean sequence length: 11 
#>    Shortest sequence: 10 
#>     Longest sequence: 12 
#> 
#> Labels:
#> one
#> two
#> 
#> Base composition:
#>     a     c     g     t 
#> 0.318 0.091 0.364 0.227 
#> (Total: 22 bases)

export_sq(sq_nuc, "Biostrings::DNAStringSet")

#>   A DNAStringSet instance of length 2
#>     width seq
#> [1]    12 TGATGAAGCGCA
#> [2]    10 TTGATGGGAA

export_sq(sq_nuc, "seqinr::SeqFastadna")

#> [[1]]
#>  [1] "T" "G" "A" "T" "G" "A" "A" "G" "C" "G" "C" "A"
#> attr(,"class")
#> [1] "SeqFastadna"
#> 
#> [[2]]
#>  [1] "T" "T" "G" "A" "T" "G" "G" "G" "A" "A"
#> attr(,"class")
#> [1] "SeqFastadna"
#>