Make your operations faster

Fast mode is meant to improve performance of operations in package. However, turning it on is associated with less control - some of parameters checks are skipped, so user has to be certain that they are correct.

Details

Not all functions support fast mode yet. Those that do not always operate in normal mode.

Turning fast mode on and off

Fast mode can be turned on and off by setting package option "tidysq_fast_mode" to accordingly TRUE and FALSE with the latter value as default. To learn more about options, see tidysq-options.

Functions supporting fast mode

Below there is a list of functions that do support fast mode:

construct_sq

read_fasta

See also