Find given motifs

Find all given motifs in all sequences and return their positions

find_motifs(sq, name, motifs)

Arguments

sq
a sq object.

name
a non-NULL character vector without NA values, containing names of the sequences in the sq. It has to be of the same length as the sq.

motifs
a character vector of motifs to be searched for.

Value

A tibble with number of rows the same as the length of sq and following columns:

name

name of the sequence

sq

sequence

sought

sought motif

found

motif found in a sequence, may differ from sought if a motif contained ambiguous letters

start

position of motif start

end

position of motif end

Details

This function allows search of a given motif or motifs in the sq object. It returns all found motifs with their start and end positions within a sequence.

Allowed and forbidden letters and characters details

Note if a sq object contains characters: ^$?=()\.|+*[] in its alphabet, search for motifs cannot be performed and an error will be displayed (with exception of sq objects of type ami - in their alphabet there is '*' letter and it can be contained in sought motif). To search for motifs with those characters, you have to replace them first using substitute_letters.

In case of sq objects of type ami and nuc, motifs have to consist of upper case letters from amino acid and nucleotide alphabet respectively. Use of lower case letters will return an error. Two additional characters are allowed: '^' and '$' indicating the beginning and the end of a sequence respectively. Moreover, notice that '*' character may be used in amino acid motifs, as it is a part of the amino acid alphabet. If a motif contains ambiguous letters, all possible matches will be searched for, e.g., amino acid motif "MAJ" (where "J" is an ambiguous letter indicating L or I) will find motifs: "MAJ", "MAL" and "MAI".

Detailed list of all letters corresponding to each ambiguous letter may be found at aminoacids_df and nucleotides_df.

See also

Examples

# Creating objects to work on:
sq_ami <- construct_sq(c("AGNTYIKFGGAYTI", "MATEGILIAADGYTWIL",
                         "MIPADHICAANGIENAGIK"), type = 'ami')
sqtbl <- read_fasta(system.file(package = "tidysq", "example_aa.fasta"),
                    type = "ami")

# Find motif of two alanines followed by aspartic acid or asparagine 
# ('AAB' motif will match 'AAB', 'AAD' and 'AAN'):
find_motifs(sq_ami, c("sq1", "sq2", "sq3"), "AAB")

#> # A tibble: 2 x 6
#>   name  sq                       sought found    start   end
#>   <chr> <(c)ami>                 <chr>  <(c)ami> <dbl> <dbl>
#> 1 sq2   MATEGILIAADGYTWIL   <17> AAB    AAD  <3>     9    11
#> 2 sq3   MIPADHICAANGIEN...  <19> AAB    AAN  <3>     9    11

# Find motif 'VHH' at the beginning of sequences:
find_motifs(sqtbl[["sq"]], sqtbl[["name"]], "^VHH")

#> # A tibble: 7 x 6
#>   name                           sq                  sought found    start   end
#>   <chr>                          <(c)ami>            <chr>  <(c)ami> <dbl> <dbl>
#> 1 AMY25|HABP3|Amyloid beta A4 p… VHHPKLVFFA...  <15> ^VHH   VHH  <3>     1     3
#> 2 AMY26|HABP4|Amyloid beta A4 p… VHHQPLVFFA...  <15> ^VHH   VHH  <3>     1     3
#> 3 AMY34|HABP12|Amyloid beta A4 … VHHQEKLVFF...  <16> ^VHH   VHH  <3>     1     3
#> 4 AMY35|HABP13|Amyloid beta A4 … VHHQEKLVFF...  <16> ^VHH   VHH  <3>     1     3
#> 5 AMY36|HABP14|Amyloid beta A4 … VHHQEKLVFF...  <16> ^VHH   VHH  <3>     1     3
#> 6 AMY38|HABP16|Amyloid beta A4 … VHHQKLVFFA...  <15> ^VHH   VHH  <3>     1     3
#> 7 AMY43|AB5|Amyloid beta A4 pep… VHHQKLVFFAEDV  <13> ^VHH   VHH  <3>     1     3

# Find motif 'DPGS' at the end of sequences:
find_motifs(sqtbl[["sq"]], sqtbl[["name"]], "DPGS$")

#> # A tibble: 1 x 6
#>   name                          sq                  sought found     start   end
#>   <chr>                         <(c)ami>            <chr>  <(c)ami>  <dbl> <dbl>
#> 1 AMY35|HABP13|Amyloid beta A4… VHHQEKLVFF...  <16> DPGS$  DPGS  <4>    13    16

# Find multiple motifs:
find_motifs(sqtbl[["sq"]], sqtbl[["name"]], c("^LIV", "XXKK", "EN$"))

#> # A tibble: 14 x 6
#>    name                         sq                  sought found     start   end
#>    <chr>                        <(c)ami>            <chr>  <(c)ami>  <dbl> <dbl>
#>  1 AMY835|Untitled|De novo      LIVAAD          <6> ^LIV   LIV   <3>     1     3
#>  2 AMY836|Untitled|De novo      LIVAGD          <6> ^LIV   LIV   <3>     1     3
#>  3 AMY837|Untitled|De novo      LIVAGE          <6> ^LIV   LIV   <3>     1     3
#>  4 AMY838|Untitled|De novo      LIVAGK          <6> ^LIV   LIV   <3>     1     3
#>  5 AMY839|Untitled|De novo      LIVAGS          <6> ^LIV   LIV   <3>     1     3
#>  6 AMY840|Untitled|De novo      LIVAGT          <6> ^LIV   LIV   <3>     1     3
#>  7 AMY95|C helix|Myohemerithrin EVVPHKKMHK...  <19> XXKK   PHKK  <4>     4     7
#>  8 AMY176|PL B1 114-138 pH 2.4… KGTFEKATSE...  <25> XXKK   TLKK  <4>    17    20
#>  9 AMY456|Untitled|Lysozyme (h… GSTDYGILQI...  <53> XXKK   CAKK  <4>    46    49
#> 10 AMY464|Untitled|Alpha synuc… GKTKEGVLYV...  <79> XXKK   FVKK  <4>    64    67
#> 11 AMY1382|Untitled|ABri        EASNCFAIRH...  <34> XXKK   TVKK  <4>    25    28
#> 12 AMY1383|A helix|Myohemerith… AM-YEQLDEE...  <24> XXKK   EHKK  <4>    10    13
#> 13 AMY392|Untitled|Insulin      LYQLEN          <6> EN$    EN    <2>     5     6
#> 14 AMY1382|Untitled|ABri        EASNCFAIRH...  <34> EN$    EN    <2>    33    34