Find elements which are not suitable for specified type.

Function finds elements in given sequence not matching to amino acid or nucleotide alphabet.

get_invalid_letters(sq, dest_type)

Arguments

sq
a sq object to be checked.

dest_type
a character string denoting destination type - it may be "nuc" for nuc type (nucleotides) or "ami" for ami type (amino acids).

Value

A list of mismatched elements for every sequence from sq object.

Details

This function allows obtaining list of mismatched elements of sequences from a sq object to amino acid or nucleotide alphabet. Output list has number of elements equal to length of sq object and each element is a character vector of elements that appear in according sequence that does not fit destination type. This function might be used to find specifically which sequences have which letters - user may want to use this information for example to check input sequences.

You can check which letters are valid for specified type in sq class documentation.

Mismatched elements might be replaced with other letters or NA using substitute_letters and then, after removal of all non-standard letters, set type to destinated using typify.

Returned lists for ami and nuc sq objects, if des_type is specified respectively "ami" and "nuc", will contain only NULL elements.

See also

Examples

# Creating an object to work on:       
sq_nucleotides <- construct_sq(c("ACGPOIUATTAGACG","GGATFGHA"))
sq_amino_acids <- construct_sq(c("QWERTYUIZXCVBNM","LKJHGFDSAZXCVBN"))

# Creating lists of mismatched elements from nucleotide sq object:
get_invalid_letters(sq_nucleotides, "nuc")

#> [[1]]
#> [1] "P" "O" "I"
#> 
#> [[2]]
#> [1] "F"
#> 

# Creating lists of mismatched elements from amino acid sq object:
get_invalid_letters(sq_amino_acids, "ami")

#> [[1]]
#> character(0)
#> 
#> [[2]]
#> character(0)
#>