Get alphabet of given sq object.

Function returns amino acid, nucleotide or atypical alphabet based on a sq object type.

get_sq_alphabet(sq)

Arguments

sq
a sq object to be recognized.

Value

A character vector of letters of the alphabet.

Details

This function allows returning alphabet of sq object which is a character or numeric vector. The function reads provided sq object and determines which kind of sequences user assigned to a sq object (nucleotide, amino acid, atypical or encoded one).

If sq type is ami the function returns a set of 20 aminoacids with gap (-) and stop codon (*) letter. If a sq contains nuc sequences the function returns a set of 5 nucleotides with gap (-) element. If sq has additionally cln subtype, ambiguous letters are returned as well.

If sq type is unt or atp the function returns a list of letters present in sequences in sq object.

If type is enc a numeric vector of values encoded for letters will be returned (see encode).

The details about amino acid and nucleotide alphabet can be checked in aminoacids_df and nucleotides_df respectively. General informations about alphabets and types of sq objects can be found in sq class documentation.

See also

Examples

# Creating an object to work on:
sq_nucleotides <- construct_sq(c("ACGATTAGACG","GGATA"), type = "nuc")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")
sq_untyped <- construct_sq(c("ACGA&&TTAGACG&"), type = "unt")

# Testing nucleotide sq object with defined type:
get_sq_alphabet(sq_nucleotides)

#> [1] "A" "C" "G" "T" "U" "-"

# Testing amino acid sq object with defined type:
get_sq_alphabet(sq_amino_acids)

#>  [1] "A" "C" "D" "E" "F" "G" "H" "I" "K" "L" "M" "N" "P" "Q" "R" "S" "T" "V" "W"
#> [20] "Y" "-" "*"

# Testing nucleotide sq object without defined type:
get_sq_alphabet(sq_untyped)

#> [1] "A" "C" "G" "&" "T"