Get type of a sq object

Function checks which type of sequences are contained in sq object.

get_sq_type(sq)

Arguments

sq
a sq object to be checked.

Value

A character string, type ofsq object - can be one of "ami", "dna", "rna", "unt", "atp" or "enc".

Details

This function returns type of sequence from sq object. If the type of sequence is dna, rna, ami, unt, atp or enc function returns "dna", "rna", "ami", "unt", "atp" or "enc" respectivetly.

See also

Examples

# Creating an object to work on:
sq_dna <- construct_sq(c("ACGATTAGACG","GGATA"), type = "dna")
sq_rna <- construct_sq(c("CCUUACGGC","UUCAAAGCU"), type = "rna")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")
sq_unt <- construct_sq(c("MMVTAAVXX"), type = "unt")
sq_atp <- substitute_letters(sq_amino_acids, c(M = "g1", V = "g2", T = "g1", A = "g3"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

sq_encoded <- encode(sq_dna, c(A = 2.3, C = 1.56, T = 0.23, G = 0.28))

# Getting sq type from DNA sq object:
get_sq_type(sq_dna)

#> dnasq 
#> "dna" 

# Getting sq type from RNA sq object:
get_sq_type(sq_rna)

#> rnasq 
#> "rna" 

# Getting sq type from amino acid  sq object:
get_sq_type(sq_amino_acids)

#> amisq 
#> "ami" 

# Getting sq type from untyped sq object:
get_sq_type(sq_unt)

#> untsq 
#> "unt" 

# Getting sq type from atypical sq object:
get_sq_type(sq_atp)

#> atpsq 
#> "atp" 

# Getting sq type from encoded sq object:
get_sq_type(sq_encoded)

#> encsq 
#> "enc"