Obtain current state of tidysq options

Subsets all global options to display those related to tidysq package.

get_tidysq_options()

Value

A named list with selected option values.

Details

The user can display value of selected option by calling getOptions(option_name) and set its value with options(option_name = value), where option_name is an option name and value is a value to assign to an option.

Full list of options included in tidysq package is listed below:

tidysq_NA_letter [character(1)]
A letter to be used when printing, constructing or interpreting NA value. Defaults to "!".

tidysq_on_warning ["none" || "message" || "warning" || "error"]
Determines the method of handling warning message. Setting "error" makes any warning throw an exception and stop execution of the code. The difference between "message" and "warning" is that while both display warning text to the console, only the latter registers it so that it can be accessed with a call to warnings(). Lastly, "none" setting causes any warnings to be completely ignored. Default value is "warning".

tidysq_pillar_max_width [codeinteger(1)]
Determines max width of a column of sq class within a tibble. Default value is 15.

tidysq_print_max_sequences [integer(1)]
Controls maximum number of sequences printed to console. If an sq object is longer than this value, then only first tidysq_print_max_sequences are printed, just like in any R vector. Default value is 10.

tidysq_print_use_color [logical(1)]
Determines whether coloring should be used to increase readability of text printed to console. While it is advised to keep this option turned on due to above concern, some environments may not support coloring and thus turning it off can be necessary. Defaults to TRUE.

tidysq_safe_mode [logical(1)]
Default value is FALSE. When turned on, safe mode guarantees that NA appears within a sequence if and only if input sequence contains value passed with NA_letter. This means that resulting type might be different to the one passed as argument, if there are letters in a sequence that does not appear in the original alphabet.

See also