Check if object has specified type

Function to checks if object is a sq object without specifying type or if it is a sq object with specific type.

is.sq(x)

is.amisq(x)

is.nucsq(x)

is.untsq(x)

is.atpsq(x)

is.encsq(x)

Arguments

x
an object to be checked.

Value

A logical value - TRUE if x has given type, FALSE otherwise.

Details

These function does not only check objects classes - they also check if their format is correct (e.g. if they have alphabet parameter, if they have exactly one type, if they are list of raws, etc. - to see details, how does sq object look like under the hood, read sq class manual).

See also

Examples

sq_ami <- construct_sq(c("CVMPQGQQ", "AHLC--PPQ"))
sq_nuc <- construct_sq(c("CGAUUACG", "UUCUAGA", "UUCA"))
sq_unt <- construct_sq("BAHHAJJ&HAN&JD&")
sq_atp <- construct_sq(c("mALPVQAmAmA", "mAmAPQ"), non_standard = "mA")
sq_enc <- encode(sq_nuc, c(A = 1.23, C = -0.72, G = 0.97, U = 3.01))

is.sq(sq_ami)

#> [1] TRUE

is.sq(sq_nuc)

#> [1] TRUE

is.sq(sq_unt)

#> [1] TRUE

is.sq(sq_atp)

#> [1] TRUE

is.sq(sq_enc)

#> [1] TRUE

is.sq(c(1,2,3))

#> [1] FALSE

is.sq(LETTERS)

#> [1] FALSE

is.sq(TRUE)

#> [1] FALSE

is.sq(NULL)

#> [1] FALSE

is.amisq(sq_ami)

#> [1] TRUE

is.nucsq(sq_nuc)

#> [1] TRUE

is.atpsq(sq_atp)

#> [1] TRUE

is.untsq(sq_unt)

#> [1] TRUE

is.encsq(sq_enc)

#> [1] TRUE

is.nucsq(sq_enc)

#> [1] FALSE

is.amisq(sq_atp)

#> [1] FALSE

is.untsq(sq_ami)

#> [1] FALSE