Get lengths of sequences in sq object

Function counts number of elements in each sequence in given sq object.

lengths.sq(x, use.names = TRUE)

# S4 method for sq
lengths(x, use.names = TRUE)

Arguments

x
an sq object.

use.names
unused argument, a leftover from default lengths method.

Value

A numeric vector, where each element gives length of according sequence from sq object.

Details

This function allows returning numeric vector of lengths of sequences from sq object. The numeric vector is as long as number of sequences present in sq object. The function counts elements in all types of sequences.

See also

Examples

# Creating an object to work on:
sq_dna <- construct_sq(c("ACGATTAGACG","GGATA"), type = "dna")
sq_amino_acids <- construct_sq(c("MMVTAAV"), type = "ami")

# Counting number of elements in DNA sq object with defined type:
lengths(sq_dna)

#> [1] 5 2

# Counting number of elements in amino acid sq object with defined type:
lengths(sq_amino_acids)

#> [1] 5