Nucleotides abbreviations

A dataset containing the nucleotide letter code and full names of nucleotides according to IUPAC nomenclature.

Format

A data frame with 17 rows and 5 columns:

one

One-letter codes of nucleotides

full

Full name of the nucleotide

amb

Logical indicating if abbreviation is ambiguous, i. e., matches more than one nucleotide

dna

Logical indicating if this nucleotide appears in standard DNA code

rna

Logical indicating if this nucleotide appears in standard RNA code

Source

Johnson, A.D. (2010). An extended IUPAC nomenclature code for polymorphic nucleic acids. Bioinformatics 26, 1386–1389.

Details

The dataset contains a nucleotide alphabet of one-letter codes and full names of the nucleotides. It also includes a gap symbol '-' that can be found in sequence alignments.

Examples

data(nucleotides_df)