Print sq object

Prints input sq object in a human-friendly form.

# S3 method for sq
print(x,
  max_sequences = getOption("tidysq_max_print_sequences"),
  use_color = getOption("tidysq_colorful_sq_print"),
  letters_sep = NULL, ...)

Arguments

x
sq object

max_sequences
numeric value indicating how many sequences should be printed

use_color
logical value indicating if sequences should be colored

letters_sep
character value indicating how the letters should be separated

...
further arguments passed to or from other methods. Unused.

Details

Print method is used by default in each case of calling the sq object with default parameters. Only by explicit calling the print method parameters can be changed.

Print checks if the input sq object is cleaned and includes this information alongside with type in the printed message. On the right side of the sequence, in angle brackets, the length of each sequence is printed (e.q. "<9>").

If the max_sequences parameter is supplied, the desired number of sequences is printed and this information is included in message (e.q. "printed 1 out of 3"). Only max_sequences value smaller then the number of sequences in object affects the function. The default value indicating how many sequences should be printed is 10, but it can be changed in package options.

Default value of use_color parameter is TRUE - sequences are printed in green and empty sequences, NA character and dots in grey. If this option is disabled, all sequences are in default color of console.

The letters_sep parameter indicates how the letters should be separated (they are not by default). Any character string can be supplied but NA_character_.

If sequences are too long, only leading characters are printed (as many as possible in single line) and following dots indicating that sequence is trunctated.

If sequences contain NA (‘Not Available’ / Missing Values) values, they are printed as "!" character, but it can be changed in package options.

This is overloaded function from base package. It is selected when sq object is used as a parameter for print function. To see the generic function page, check here.

See also

Examples

# Creating sq objects using construct_sq:
sq_ami <- construct_sq(c("MIAANYTWIL","TIAALGNIIYRAIE",
                         "NYERTGHLI", "MAYXXXIALN"), type = "ami")
sq_nuc <- construct_sq(c("ATGCAGGA", "GACCGAACGAN",
                         "TGACGAGCTTA"), type = "nuc")
sq_unt <- construct_sq(c("ATGCAGGA!", "TGACGAGCTTA", "", "TIAALGNIIYRAIE"))

# Printing without explicit function calling with default parameters:
sq_ami

#> ami (amino acids) sequences list:
#> [1] MIAANYTWIL                                                              <10>
#> [2] TIAALGNIIYRAIE                                                          <14>
#> [3] NYERTGHLI                                                                <9>
#> [4] MAYXXXIALN                                                              <10>
#> 

sq_nuc

#> nuc (nucleotides) sequences list:
#> [1] ATGCAGGA                                                                 <8>
#> [2] GACCGAACGAN                                                             <11>
#> [3] TGACGAGCTTA                                                             <11>
#> 

sq_unt

#> unt (unspecified type) sequences list:
#> [1] ATGCAGGA!                                                                <9>
#> [2] TGACGAGCTTA                                                             <11>
#> [3] <NULL>                                                                   <0>
#> [4] TIAALGNIIYRAIE                                                          <14>
#> 

# Printing with explicit function calling and specific parameters:
print(sq_ami)

#> ami (amino acids) sequences list:
#> [1] MIAANYTWIL                                                              <10>
#> [2] TIAALGNIIYRAIE                                                          <14>
#> [3] NYERTGHLI                                                                <9>
#> [4] MAYXXXIALN                                                              <10>
#> 

print(sq_nuc, max_sequences = 1, use_color = FALSE)

#> nuc (nucleotides) sequences list:
#> [1] ATGCAGGA                                                                 <8>
#> printed 1 out of 3

print(sq_unt, letters_sep = ":")

#> unt (unspecified type) sequences list:
#> [1] A:T:G:C:A:G:G:A:!                                                        <9>
#> [2] T:G:A:C:G:A:G:C:T:T:A                                                   <11>
#> [3] <NULL>                                                                   <0>
#> [4] T:I:A:A:L:G:N:I:I:Y:R:A:I:E                                             <14>
#> 

# Printing of the cleaned object:
clean(sq_nuc)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] ATGCAGGA                                                                 <8>
#> [2] <NULL>                                                                   <0>
#> [3] TGACGAGCTTA                                                             <11>
#> 

print(clean(sq_nuc), letters_sep = "-", use_color = FALSE)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] A-T-G-C-A-G-G-A                                                          <8>
#> [2] <NULL>                                                                   <0>
#> [3] T-G-A-C-G-A-G-C-T-T-A                                                   <11>
#>