Generate random sequences

Generates a sq object with specified number of sequences of given length and given type.

random_sq(n, len, type, is_clean, sd = NULL, use_gap = FALSE)

Arguments

n
a positive integer value - number of sequences to generate.

len
a positive integer value - length of each sequence if sd not specified and mean length of sequences if sd specified

type
a type of generated sq object; possible values are "ami", "dna" and "rna" (see section sq types in sq documentation for details).

is_clean
a logical value - if TRUE, letters will be drawn from clean alphabet, if FALSE - ambiguous letters might be also generated.

sd
a positive numeric value; if specified, gives standard deviation of length of generated sequences.

use_gap
- a logical value; if TRUE, sequences will be generated with random gaps inside.

Value

An object of class sq with type as specified.

Sequences are generated using sample function. There is no possibility of generating a sequence of length 0, even if sd is given. Letter '*' is not used in generating ami sequences.

See also

Examples

# setting seed for reproducibility
set.seed(16)

# generating random sequences
random_sq(10, 10, "ami", TRUE)

#> ami (amino acids), cln (cleaned) sequences list:

#>  [1] QFLFVHCTVD                                                             <10>
#>  [2] HNEQMFTQRC                                                             <10>
#>  [3] YTMKFPYYSL                                                             <10>
#>  [4] AVRVQLGVVH                                                             <10>
#>  [5] AKICKSWGES                                                             <10>
#>  [6] WMVHTPSSNT                                                             <10>
#>  [7] GYNTLDELEC                                                             <10>
#>  [8] PKRLNCPIIE                                                             <10>
#>  [9] CNYWNWIDGE                                                             <10>
#> [10] PTMPRGRLLY                                                             <10>

random_sq(25, 18, "rna", TRUE, sd = 6)

#> rna (RNA), cln (cleaned) sequences list:

#>  [1] GUCUAUCUUCGCACU                                                        <15>
#>  [2] UACGUAAUCCGGAGGUGUAAG                                                  <21>
#>  [3] GGGAGUAUAGCUCGUGAAGCG                                                  <21>
#>  [4] CGUGUCGGCACGAUC                                                        <15>
#>  [5] CCAACAUUUUCCGC                                                         <14>
#>  [6] UACUCAAGUCUGUUGCGGUCAGACACU                                            <27>
#>  [7] AUCCCACGAAAGAAGUGAGG                                                   <20>
#>  [8] ACGUCAGGGCAGACGGCUCC                                                   <20>
#>  [9] CUCACAUGCGUACAAUUGGC                                                   <20>
#> [10] UGAGUUGCGU                                                             <10>

#> printed 10 out of 25

random_sq(50, 8, "dna", FALSE, sd = 3)

#> dna (DNA) sequences list:

#>  [1] VCRCSGYW                                                                <8>
#>  [2] TCGWKKGKTV                                                             <10>
#>  [3] TNHDDSD                                                                 <7>
#>  [4] TYWVAVTWWRC                                                            <11>
#>  [5] VGKBHBHDSVWRH                                                          <13>
#>  [6] YG                                                                      <2>
#>  [7] AMBWW                                                                   <5>
#>  [8] YMSAS                                                                   <5>
#>  [9] YWARKSC                                                                 <7>
#> [10] HMRRVTARK                                                               <9>

#> printed 10 out of 50

random_sq(6, 100, "ami", TRUE, use_gap = TRUE)

#> ami (amino acids), cln (cleaned) sequences list:

#> [1] GHK-QFPESSVFFMSEPIAAWCRMYYWFWHSPIIHRVPYWKWHKS-TRCLMPPVKAQQCWA-FCQAL... <100>
#> [2] SP-GGRAFPFVVMWHHVNSVHCYDWSHVTEC-YHDELIYSQESRLEQMQLPYIIWELLTK-REVSQH... <100>
#> [3] DQGQQR-GGISHHDMHNYNNVFWIRWLHVNGPMLTPK-HVWLCPMWWAMEQGP-DNRLNCQHIPNYP... <100>
#> [4] MQMKMMQCKREVIGLHQAFQMMQVHWCKKCP-EIMPHYWVICVQQCK-NQTVITICCNEAGHYTSCL... <100>
#> [5] KL-P-CPRIDTPGKF-SHYDP---WKMFFYINPVQWDWCQRPRMMVHWRPECYFVRLREASPAGYTM... <100>
#> [6] WNGILTAQCQEEIYKRWTGHQGRMHC-MWVHKGTFIQPNWKGGHWPPWTSLDCSQWYCQC-DMPVTH... <100>