Read a FASTA file

Reads a FASTA file of nucleotides or amino acids file and returns a tibble with number of rows corresponding to the number of sequences and two columns: 'name' and 'sq' giving the name of the sequence and the sequence itself.

read_fasta(file, type = NULL, is_clean = NULL, non_standard = NULL)

read_fasta_ami(file, is_clean = NULL)

read_fasta_nuc(file, is_clean = NULL)

Arguments

file
a character string indicating path to file or url.

type
a character string indicating type of sq object that is going to be constructed; supported values are "ami" for amino acid sequences, "nuc" for nucleotide sequences, "unt" for and NULL for type guessing (see details)

is_clean
a logical value indicating if sequences are clean. or in other words - they don't contain ambiguous values; supported values are TRUE for clean sequences, FALSE for unclean sequences and NULL for auto detecting (see details).

non_standard
a character vector indicating non-standard letters contained in sequences. If NULL, sequences will not be searched for non-standard letters of length more than one. Each element of this parameter should be at least two characters long.

Value

A tibble with number of rows corresponding to the number of sequences and two columns: 'name' and 'sq' giving the name of the sequence and the sequence itself.

Details

All rules of creating sq objects are the same as in construct_sq.

Functions read_fasta_ami and read_fasta_nuc are wrappers around read_fasta with specified type parameter - accordingly "ami" or "nuc". You can also pass "is_clean" parameter to those functions, but you cannot pass "non_standard".

See also

Examples

fasta_file <- system.file(package = "tidysq",
                     "sample_fasta/sample_ami.fasta")
read_fasta(fasta_file)

#> # A tibble: 100 x 2
#>    name          sq                      
#>    <chr>         <ami>                   
#>  1 sample_ami 1  KWOVPFWSAZFBTPS...  <29>
#>  2 sample_ami 2  EDAJZOIFQLVTAYV...  <32>
#>  3 sample_ami 3  GXAYZJKAEOMETFG...  <28>
#>  4 sample_ami 4  CCTKOSTJAEDGMIK...  <27>
#>  5 sample_ami 5  DCRTRZJWBHRPHXH...  <30>
#>  6 sample_ami 6  PSQLECWRPRGQZFP...  <29>
#>  7 sample_ami 7  DKWLELEWMYCWQJY...  <28>
#>  8 sample_ami 8  LLBOMGZSKZIEXDP...  <28>
#>  9 sample_ami 9  KAZBIBGOMLLDYHH...  <25>
#> 10 sample_ami 10 FZFTAMEOKKIMZUF...  <28>
#> # … with 90 more rows

read_fasta_ami(fasta_file)

#> # A tibble: 100 x 2
#>    name          sq                      
#>    <chr>         <ami>                   
#>  1 sample_ami 1  KWOVPFWSAZFBTPS...  <29>
#>  2 sample_ami 2  EDAJZOIFQLVTAYV...  <32>
#>  3 sample_ami 3  GXAYZJKAEOMETFG...  <28>
#>  4 sample_ami 4  CCTKOSTJAEDGMIK...  <27>
#>  5 sample_ami 5  DCRTRZJWBHRPHXH...  <30>
#>  6 sample_ami 6  PSQLECWRPRGQZFP...  <29>
#>  7 sample_ami 7  DKWLELEWMYCWQJY...  <28>
#>  8 sample_ami 8  LLBOMGZSKZIEXDP...  <28>
#>  9 sample_ami 9  KAZBIBGOMLLDYHH...  <25>
#> 10 sample_ami 10 FZFTAMEOKKIMZUF...  <28>
#> # … with 90 more rows
# NOT RUN {
read_fasta("https://www.uniprot.org/uniprot/P28307.fasta")
# }