sq: class for keeping biological sequences tidy

An object of class sq represents a list of biological sequences. It is the main internal format of the tidysq package and most functions operate on it. The storage method is memory-optimized so that objects require as little memory as possible (details below).

Construction/reading/import of sq objects

There are multiple ways of obtaining sq objects:

constructing from a character vector with construct_sq,

constructing from another object with as.sq method,

reading from the fasta file with read_fasta,

importing from a format of other package like ape or Biostrings with import_sq.

Important note: A manual assignment of a class sq to an object is strongly discouraged - due to the usage of low-level functions for bit packing such assignment may lead to calling one of those functions during operating on object or even printing it which can cause a crash of R session and, in consequence, loss of data.

Export/writing of sq objects

There are multiple ways of saving sq objects or converting them into other formats:

converting into a character vector with as.character method,

converting into a character matrix (or numeric matrix, in case of enc) with as.matrix method,

converting into a list of numerics with encsq_to_list (only for enc sq),

saving as fasta file with write_fasta,

exporting into a format of other package like ape or Biostrings with export_sq.

Types of sq

This package is meant to handle both amino acids and nucleotides sequences thus there is need to differentiate sq objects that keep them. In addition, there are special types for handling non-standard sequence formats and encodings.

Each sq object has exactly one of types:

ami - (amino acids) represents a list of sequences of amino acids (peptides or proteins),

nuc - (nucleotides) represents a list of nucleotide sequences (RNA or DNA),

unt - (untyped) represents a list of sequences that do not have specified type. They are mainly result of reading sequences from a file that contains some letters that are not in standard nucleotide or amino acid alphabets and user has not specified them explicitly. They should be converted to ami or nuc sequences (using functions like substitute_letters or typify).

atp - (atypical) represents sequences that have an alphabet different from standard ami or nuc alphabets - similarly to unt, but user has explicitly informed about it. They are result of constructing sequences or reading from file with specifying non_standard parameter (for details see read_fasta and construct_sq). They are also result of using function substitute_letters - users can use this for example to simplify an alphabet and to replace a few letters by one.

enc - (encoded) represents list of sequences that have been encoded with function encode where each letter is assigned with numeric value.

Additionally, there is a special subtype cln (standing for clean). Only ami and nuc sq objects may have this subtype. It indicates that sequences do not contain ambiguous letters (see "alphabets" section below).

sq object type is printed when using overloaded method print. It can be also checked by using get_sq_type

Alphabet

Each sq object have an alphabet associated with it. Alphabet is a set of possible letters that can appear in sequences contained in object. Alphabet is kept mostly as a character vector, where each element represents one letter.

sq objects of type ami or nuc have fixed alphabets (depending also if object has cln subtype). In other words, if two sq objects have exactly the same type - ami or nuc - and either both have or both do not have a cln subtype, they are ensured to have the same alphabets.

Here are listed alphabets for these types:

ami cln - ACDEFGHIKLMNPQRSTVWY-*

ami, (not cln) - ABCDEFGHIJKLMNOPQRSTUVWXYZ-*

nuc cln - ACGTU-

nuc, (not cln) - ACGTUWSMKRYBDHVN-

To see details of these alphabets see aminoacids_df and nucleotides_df.

Other types of sq objects are allowed to have different alphabets. Having an alphabet exactly identical to one of those above does not automatically indicate that the type of the sequence is one of those - e. g., there might be unt sq that has an alphabet identical to ami cln alphabet. To set the type, one should use the typify function.

The purpose of co-existence of unt and atp alphabets is the fact that although there is a standard for format of fasta files, sometimes there are other types of symbols, which do not match the standard. Thanks to these types, tidysq can import files with customized alphabets can be imported. Moreover, a user may want to group amino acids with similar properties (e. g., for machine learning) and replace the longer alphabet with symbols of groups. To check details, see read_fasta, construct_sq and substitute_letters.

All of the types: ami, nuc, atp, unt have alphabets that are character vectors while enc objects have alphabets that are numeric vectors. They are result of encode function, see its manual for details.

Important note: in atp alphabets there is possibility of appearance of letters that consists of more than one character - this functionality is provided in order to handle situations like post-translational modifications, (e.g., using "mA" to indicating methylated alanine).

Important note: alphabets of atp and unt sq objects are case sensitive. Thus, in their alphabets there can appear both lowercase and uppercase letters simultaneously and they are treated as different characters. Alphabets of nuc and ami objects are always uppercase and all functions converts other parameters to uppercase when working with ami or nuc - e. g., %has% operator converts lower letters to upper when searching for motifs in ami or nuc object.

Important note: maximum length of an alphabet is 30 letters. You are not allowed to read fasta files or construct from character vectors that have more than 30 distinct characters in sequences (with exception of reading or constructing ami or nuc objects - during their construction lowercase letters are automatically converted to uppercase).

You can obtain an alphabet of the sq object using the get_sq_alphabet function. You can check which letters are invalid (are not represented in standard amino acids or nucleotides alphabet) in each sequence of given sq object of type unt or atp by using get_invalid_letters. You can substitute one letter with another using substitute_letters.

Missing/Not Available values

There is a possibility of introducing NA values into sequences. NA value does not represents gap (which are represented by '-') or wildcard elements (N in he case of nucleotides and 'X' in the case of amino acids), but is used as a representation of an empty position or invalid letters (not represented in nucleotide or amino acid alphabet). NA does not belong to any alphabet (with exception of enc objects where some of letters might be NA, see encode for details). It is printed as '!' and, thus, it is highly unrecommended to use '!' as special letter in atp sequences (but print character can be changed in options, see tidysq-options).

NA might be introduced by:

reading fasta file with non-standard letters in fast mode with read_fasta,

replacing a letter with NA value with substitute_letters,

subsetting sequences out of their lengths with bite.

An user can convert sequences that contain NA values into NULL sequences with remove_na.

NULL (empty) sequences

NULL sequence is a sequence of length 0.

NULL sequences might be introduced by:

constructing sq object from character string of length zero with construct_sq,

using the clean function,

using the remove_na function,

subsetting sq object with extract operator

Storage format

sq object is, in fact, list of raw vectors. The fact that it is list implies that an user can concatenate sq objects using c method and subset them using extract operator. Alphabet is kept as an attribute of the object.

Raw vectors are the most efficient way of storage - each letter of the sequence has assigned an integer (its index in alphabet of sq object). Those integers in binary format fit in less than 8 bits, but normally are stored on 16 bits. However, thanks to bit packing it is possible to remove unused bits and store numbers more tightly. This operations result in a little time overhead in all operations, because most of them require unpacking and repacking sequences, but this cost is relatively low in comparison to amount of saved memory.

For example - nuc cln alphabet consist of 6 values: ACGTU-. They are assigned numbers 1 to 6 respectively. Those numbers in binary format take form: 001, 010, 011, 100, 101, 110. Each of the letters can be coded with just 3 bits instead of 8 which is demanded by char - this allows us to save more than 60% of memory spent on storage of nucleotides sequences.

tibble compatibility

sq objects are compatible with tibble class - that means one can have a sq object as a column of a tibble. There are overloaded print methods, so that it is printed in pretty format.