Substitute letters in a sequence

1) Replace ambiguous/extraordinary letters in a nucleic or amino acid sequence, stored in a sq object, with the ones that are compliant with the IUPAC standard, ones that are user-defined or with NA values.

2) Replace default amino acid letters in a sequence with a custom encoding to create sequences with simplified alphabets.

The function is only used to replace letters in the alphabet. It cannot be used to merge multiple characters into one.

substitute_letters(sq, encoding)

Arguments

sq
a sq object.

encoding
a vector of letters to be replaced together with their replacements. One letter can be replaced with multiple symbols. To perform substitution create a named vector, e.g. c(A = Ala, H = His, amino_or_nucleic_acid_symbol = replacement).

Value

a sq object with atp type with replaced alphabet, defined by user.

Details

substitute_letters allows to replace ambiguous/extraordinary letters in nucleic or amino acid sequence with user-defined or IUPAC symbols. Letters can also be replaced with NA values, so that they can be later removed, from the sequence, by clean function.

substitute_letters can be used to replace default amino acid letters with encodings. They can be user-defined or be derived from various simplified alphabets.

One letter of the alphabet may be replaced by a multiple character.

The alphabet characters to be replaced need to be written in capital letters and must originate from default alphabets, otherwise error will be introduced.

Multiple string of letters to be substituted (ex. c(AHG = "replacement")) will also produce an error.

Replacing multiple letters with the same symbol (ex. c(A = "rep1", H = "rep1", G = "rep1")) is allowed.

Created sequence will be stripped of cln subtype, if the original sequence possessed it. This will also occur when the letter to be replaced will not be found in the sequence. It remain unchanged but will lose subclass.

The newly constructed will have a new type atp, representing sequences with atypical alphabet.

All replaced letters will have the character type. Multiple symbol replacement will be recognized as one letter and cannot be separated in future operations into single letters.

See also

Examples

# Creating sq object to work on:

sq_dna <- construct_sq(c("TATGAATTAGCTGTCTTTGCTGCTTTGGTTATCTATGA", 
                         "CTTTGGTTATCTAGCTGTATGA", "TATCTAGCTGTATG", 
                         "CTGCTG", "CTTAGA", "CCCT", "CTGAATGT"), 
                       type = "dna")

sq_ami <- construct_sq(c("NYMITGGREEYERTVIYRAIALNAANYTWIL", 
                        "TIAALGNIIYRAIE", "NYERTGHLI", 
                        "MAYNNNIALN", "MN", "NAAAT"), 
                        type = "ami")
                     

# Replace single letter of alphabet with single character encoding:

substitute_letters(sq_dna, c(T = "t", A = "a", C = "H", G = "Z"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] tatZaattaZHtZtHtttZHtZHtttZZttatHtatZa                                  <38>
#> [2] HtttZZttatHtaZHtZtatZa                                                  <22>
#> [3] tatHtaZHtZtatZ                                                          <14>
#> [4] HtZHtZ                                                                   <6>
#> [5] HttaZa                                                                   <6>
#> [6] HHHt                                                                     <4>
#> [7] HtZaatZt                                                                 <8>

substitute_letters(sq_dna, c(T = 1, A = 2, C = 3, G = 4))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] 12142211243141311143143111441121312142                                  <38>
#> [2] 3111441121312431412142                                                  <22>
#> [3] 12131243141214                                                          <14>
#> [4] 314314                                                                   <6>
#> [5] 311242                                                                   <6>
#> [6] 3331                                                                     <4>
#> [7] 31422141                                                                 <8>

substitute_letters(sq_ami, c(M = "m", Q = "g", R = "#", D = "$"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] NYmITGG#EEYE#TVIY#AIALNAANYTWIL                                         <31>
#> [2] TIAALGNIIY#AIE                                                          <14>
#> [3] NYE#TGHLI                                                                <9>
#> [4] mAYNNNIALN                                                              <10>
#> [5] mN                                                                       <2>
#> [6] NAAAT                                                                    <5>

substitute_letters(sq_ami, c(M = "2", Q = "5", R = "9", D = "7"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] NY2ITGG9EEYE9TVIY9AIALNAANYTWIL                                         <31>
#> [2] TIAALGNIIY9AIE                                                          <14>
#> [3] NYE9TGHLI                                                                <9>
#> [4] 2AYNNNIALN                                                              <10>
#> [5] 2N                                                                       <2>
#> [6] NAAAT                                                                    <5>

# Replace single letter of alphabet with multiple character encoding:

substitute_letters(sq_dna, c(T = "th", A = "ad", C = "cy", G = "gu"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] th ad th gu ad ad th th ad gu cy th gu th cy th th th gu cy th gu cy... <38>
#> [2] cy th th th gu gu th th ad th cy th ad gu cy th gu th ad th gu ad       <22>
#> [3] th ad th cy th ad gu cy th gu th ad th gu                               <14>
#> [4] cy th gu cy th gu                                                        <6>
#> [5] cy th th ad gu ad                                                        <6>
#> [6] cy cy cy th                                                              <4>
#> [7] cy th gu ad ad th gu th                                                  <8>

substitute_letters(sq_dna, c(T = 111, A = 222, C = 333, G = 444))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] 111 222 111 444 222 222 111 111 222 444 333 111 444 111 333 111 111...  <38>
#> [2] 333 111 111 111 444 444 111 111 222 111 333 111 222 444 333 111 444...  <22>
#> [3] 111 222 111 333 111 222 444 333 111 444 111 222 111 444                 <14>
#> [4] 333 111 444 333 111 444                                                  <6>
#> [5] 333 111 111 222 444 222                                                  <6>
#> [6] 333 333 333 111                                                          <4>
#> [7] 333 111 444 222 222 111 444 111                                          <8>

substitute_letters(sq_ami, c(M = "Met", Q = "Gln", R = "Arg", D = "Asp"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] N Y Met I T G G Arg E E Y E Arg T V I Y Arg A I A L N A A N Y T W I L   <31>
#> [2] T I A A L G N I I Y Arg A I E                                           <14>
#> [3] N Y E Arg T G H L I                                                      <9>
#> [4] Met A Y N N N I A L N                                                   <10>
#> [5] Met N                                                                    <2>
#> [6] N A A A T                                                                <5>

substitute_letters(sq_ami, c(M = "222", Q = "555", R = "999", D = "777"))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] N Y 222 I T G G 999 E E Y E 999 T V I Y 999 A I A L N A A N Y T W I L   <31>
#> [2] T I A A L G N I I Y 999 A I E                                           <14>
#> [3] N Y E 999 T G H L I                                                      <9>
#> [4] 222 A Y N N N I A L N                                                   <10>
#> [5] 222 N                                                                    <2>
#> [6] N A A A T                                                                <5>

# Replace single letter of alphabet with NA value:

substitute_letters(sq_dna, c(A = NA, G = NA))

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] T!T!!!TT!!CT!TCTTT!CT!CTTT!!TT!TCT!T!!                                  <38>
#> [2] CTTT!!TT!TCT!!CT!T!T!!                                                  <22>
#> [3] T!TCT!!CT!T!T!                                                          <14>
#> [4] CT!CT!                                                                   <6>
#> [5] CTT!!!                                                                   <6>
#> [6] CCCT                                                                     <4>
#> [7] CT!!!T!T                                                                 <8>

# Use created encoding

sub_dna <- c(T = "t", A = "a", C = "c", G = "g")
sub_ami <- c(M = "Met", Q = "Gln", R = "Arg", D = "Asp", 
             H = "His", K = "Lys", A = "Ala")

substitute_letters(sq_dna, sub_dna)

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] tatgaattagctgtctttgctgctttggttatctatga                                  <38>
#> [2] ctttggttatctagctgtatga                                                  <22>
#> [3] tatctagctgtatg                                                          <14>
#> [4] ctgctg                                                                   <6>
#> [5] cttaga                                                                   <6>
#> [6] ccct                                                                     <4>
#> [7] ctgaatgt                                                                 <8>

substitute_letters(sq_ami, sub_ami)

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] N Y Met I T G G Arg E E Y E Arg T V I Y Arg Ala I Ala L N Ala Ala N...  <31>
#> [2] T I Ala Ala L G N I I Y Arg Ala I E                                     <14>
#> [3] N Y E Arg T G His L I                                                    <9>
#> [4] Met Ala Y N N N I Ala L N                                               <10>
#> [5] Met N                                                                    <2>
#> [6] N Ala Ala Ala T                                                          <5>

# Use created encoding from other package 
# (ex. AmyloGram::AmyloGram_model)

library(AmyloGram)

AG_sub_raw <- unlist(AmyloGram_model[["enc"]])

sub_AG <- substr(names(AG_sub_raw), 1, 1)
names(sub_AG) <- toupper(AG_sub_raw)
sub_AG 

#>   G   K   P   R   I   L   V   F   W   Y   A   C   H   M   D   E   N   Q   S   T 
#> "1" "2" "2" "2" "3" "3" "3" "4" "4" "4" "5" "5" "5" "5" "6" "6" "6" "6" "6" "6" 

substitute_letters(sq_ami, sub_AG)

#> Warning: 'sq' object passed to substitute_letters had 'cln' subtype, output doesn't have it

#> atp (atypical alphabet) sequences list:
#> [1] 6453611266462633425353655646433                                         <31>
#> [2] 63553163342536                                                          <14>
#> [3] 646261533                                                                <9>
#> [4] 5546663536                                                              <10>
#> [5] 56                                                                       <2>
#> [6] 65556                                                                    <5>