Set options of package

You can get value of an option by calling getOptions(option_name) and set its value by calling options(option_name = value), where option_name is an option name (full list of this package included below) and value is a value to assing to an option.

Details

You can change default behaviour of package using one of following options:

tidysq_bite_na_action (default "warning") - a character string specifying in which way to inform user about biting sequences out of the range when using bite; possible values: "error", "warning", "message", "none",

tidysq_subsitute_letters_cln (default "warning") - a character string specifying in which way to inform user about droping cln subtype of sq while using substitute_letters; possible values: "error", "warning", "message", "none",

tidysq_typify_small_cap_let (default "warning") - a character string specifying in which way to inform user about merging lowercase and uppercase letters when typifying sq object; possible values: "error", "warning", "message", "none",

tidysq_encode_no_given_action (default "warning") - a character string specifying in which way to inform user about encoding unspecified letters as NA if they do appear in sequences in encode; possible values: "error", "warning", "message", "none",

tidysq_max_pillar_sq_width (default 15) - an integer value specyfying pillar_shaft_sq width

tidysq_max_print_sequences (default 10) - an integer value specyfying maximum number of printed sequences in print sq,

tidysq_colorful_sq_print (default TRUE) - a logical value if to use colorful printing (see print sq)

tidysq_na_print_char (default = "!") - a character string which character string to pring when NA values appear in sequences,

tidysq_fast_mode (default FALSE) - a logical value if to work in fast mode (see fast-mode)

If value is not appropriate for given option, default value will be used.

See also