Set type of untyped and atypical sequences

Function to set type of atp or unt sequences to ami, dna or rna provided that they do not have invalid letters (which can be checked using get_sq_alphabet and get_invalid_letters).

typify(sq, dest_type)

Arguments

sq
an object of class sq with one of the types ami, dna, rna, unt or atp.

dest_type
character string, destination type, one of "ami", "dna" or "rna".

Value

An object of class sq that represents the same sequences as input sq, but with type as specified in dest_type.

Details

Sometimes during reading sq from fasta file (using read_fasta) or constructing from character vector (using construct_sq or as.character) there are letters that are not in standard alphabets (see aminoacids_df and nucleotides_df). In consequence, newly created sq objects might have other type - atp or unt. After removal of those non-standard letters (using substitute_letters), user might want to set type of sq object to one of standard types - ami, dna and rna - as this is demanded by some functions. This is what this function is designed for.

If dest_type is equal to type of sq, function does not do anything.

If sq object contain letters that are not in specified destination alphabet, an error will be thrown. However, typify converts automatically lowercase letters to uppercase letters, so e.g. "a" is treated as element of amino acid alphabet.

If sq contains both lower and uppercase letters, they will be converted to uppercase, but a message informing about it will be printed in the console. This action is default and can be changed in package options (see tidysq-options).

Output sq object will not have cln subtype, even if all letters of it fit in clean alphabet of destination type (with exception of passing already clean object as input).

Examples

# constructing artificial sq object with strange characters - type will be "unt"
sq_unt <- construct_sq(c("&VPLG&#", "##LCG"))

# substituting letters with "X", which stands for unknown amino acid
sq_sub <- substitute_letters(sq_unt, c(`&` = "X", `#` = "X"))

# setting type
typify(sq_sub, "ami")

#> ami (amino acids) sequences list:

#> [1] XVPLGXX                                                                  <7>
#> [2] XXLCG                                                                    <5>