Convert sq object into matrix

Coerce sequences from a sq object to a matrix, in which rows correspond to sequences and columns to positions

# S3 method for sq
as.matrix(x, ...)

Arguments

x
a sq object to be converted.

...
further arguments to be passed from or to other methods.

Value

A matrix with number of rows the same as number of sequences and number of columns corresponding to the length of the longest sequence in the converted sq object. Matrix is either character (if type of sq is ami, nuc, atp or unt) or numeric (if type is enc).

Details

This method for class sq allows converting sequences from the sq object into a matrix. Each row corresponds to the seperate sequence from the sq object, whereas each column indicates a single position within a sequence. Dimensions of matrix are determined by the number of sequences (rows) and the length of the longest sequence (columns). If a length of sequence is smaller than the lenght of the longest sequence, the remaining columns will be filled with NA. All attributes of the input sq are lost during the conversion to matrix.

See also

Examples

# Creating objects to work on:
sq_same_len <- construct_sq(c("CGATAGACA", "TGACAAAAC", "GTGACCGTA"),
                            type = 'nuc')
sq_diff_len <- construct_sq(c("CTGAATGCAGTACCGTAAT", "ATGCCGTAAATGCCAT",
                              "CAGACCANNNATAG"), type = 'nuc')

# Converting sq object containing sequences of the same lengths to matrix:
as.matrix(sq_same_len)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9]
#> [1,] "C"  "G"  "A"  "T"  "A"  "G"  "A"  "C"  "A" 
#> [2,] "T"  "G"  "A"  "C"  "A"  "A"  "A"  "A"  "C" 
#> [3,] "G"  "T"  "G"  "A"  "C"  "C"  "G"  "T"  "A" 

# Converting sq object containing sequences of different lengths to 
# matrix - NA introduced:
as.matrix(sq_diff_len)

#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13] [,14]
#> [1,] "C"  "T"  "G"  "A"  "A"  "T"  "G"  "C"  "A"  "G"   "T"   "A"   "C"   "C"  
#> [2,] "A"  "T"  "G"  "C"  "C"  "G"  "T"  "A"  "A"  "A"   "T"   "G"   "C"   "C"  
#> [3,] "C"  "A"  "G"  "A"  "C"  "C"  "A"  "N"  "N"  "N"   "A"   "T"   "A"   "G"  
#>      [,15] [,16] [,17] [,18] [,19]
#> [1,] "G"   "T"   "A"   "A"   "T"  
#> [2,] "A"   "T"   NA    NA    NA   
#> [3,] NA    NA    NA    NA    NA