Subset sequences from sq objects

Extracts a defined range of elements (amino acids or nucleotides) from a sequence.

bite(sq, indices)

Arguments

sq
a sq object.

indices
a numeric vector of subsequence indices to extract from each sequence. The function follows the normal R conventions for indexing vectors, including negative indices.

Value

sq object of the same type as input sq, where each element is a subsequence created by indexing corresponding sequence from input sq object with input indices.

Details

Amino acids and nucleic acid sequences are represented as sq object in the tidysq package. Often one needs to get only a single letter or the sequence of a defined range from the original sequences. A subsequence is a sequence that can be derived from the original sequence by trimming some elements (letters) without changing the order of the remaining elements. To obtain a subsequence from each sequence contained in the sq object with the same indices. This is for example useful to extract an user-defined region from a sequence.

The usage of bite follows the normal R conventions. For details refer to the R documentation (see https://cran.r-project.org/doc/manuals/r-release/R-intro.html#Index-vectors). Extracting indices not present in the sequence results in introducing NA (‘Not Available’ / Missing Values) values. Information about it is printed on console depending on value of option 'tidysq_bite_na_action' - it can be either a warning (default), error, message or no information (you can check details in tidysq-options). NA values can be removed by using remove_na function.

See also

Examples

# Creating object, called sq to work on:
# The first four sequences of are miRNAs and isomiRs that were taken from 
# Tan et al. (2014) Nucleic Acid Research, doi: 10.1093/nar/gku656. 
# Selected miR-9 (isomiRs) sequences from neural progenitor stem cells
# look as follows: TCTTTGGTTATCTAGCTGTATGA, CTTTGGTTATCTAGCTGTATGA, 
# TCTTTGGTTATCTAGCTGTATG, TCTTTGGTTATCTAGCTGTATGAA
# The remaining sequences represent short artificial random sequences.
# Tan et al. have shown that "majority of miRNA genes encode mature isomers 
# that vary in size by one or more bases at the 3' and/or 5' end of the miRNA.
# In the following the bite function is used to trim nucleotieds 5' for a 
# simple visual alignment.
# construct_sq is use to create an sq object

sq <- construct_sq(c("TCTTTGGTTATCTAGCTGTATGA", "CTTTGGTTATCTAGCTGTATGA",
                     "TCTTTGGTTATCTAGCTGTATG", "TCTTTGGTTATCTAGCTGTATGAA",
                     "ACTGCTG", "CTTAGA", "CCCT", "CTGAATGT"), type = "nuc")

# Get an overview of the sequences and show the first four only:
# The first four isomiRs sequences have lengths of 15, 14, 14 and 15
# nucleotides. The remaining artificial sequences have lengths of 5, 4, 3 
# and 5 nucleotides, respectively.
summary(sq)

#>      Length Class  Mode
#> [1,] 9      -none- raw 
#> [2,] 9      -none- raw 
#> [3,] 9      -none- raw 
#> [4,] 9      -none- raw 
#> [5,] 3      -none- raw 
#> [6,] 3      -none- raw 
#> [7,] 2      -none- raw 
#> [8,] 3      -none- raw 

# Working with the first four miRNA/isomiR
# Removing first letter from the first four sequences (miRNAs/isomiRs).
# The sequence 1, 3 and 4 appear to be more similar.
bite(sq[1:4], -1)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] CTTTGGTTATCTAGCTGTATGA                                                  <22>
#> [2] TTTGGTTATCTAGCTGTATGA                                                   <21>
#> [3] CTTTGGTTATCTAGCTGTATG                                                   <21>
#> [4] CTTTGGTTATCTAGCTGTATGAA                                                 <23>
#> 

# Extracting first five letters from each miRNA/ismomiR sequence:
bite(sq[1:4], 1:5)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TCTTT                                                                    <5>
#> [2] CTTTG                                                                    <5>
#> [3] TCTTT                                                                    <5>
#> [4] TCTTT                                                                    <5>
#> 

# Working with all sequences
# extracting first letter from each sequence:
bite(sq, 1)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] T                                                                        <1>
#> [2] C                                                                        <1>
#> [3] T                                                                        <1>
#> [4] T                                                                        <1>
#> [5] A                                                                        <1>
#> [6] C                                                                        <1>
#> [7] C                                                                        <1>
#> [8] C                                                                        <1>
#> 

# Extracting first three letters from each sequence:
bite(sq, 1:3)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TCT                                                                      <3>
#> [2] CTT                                                                      <3>
#> [3] TCT                                                                      <3>
#> [4] TCT                                                                      <3>
#> [5] ACT                                                                      <3>
#> [6] CTT                                                                      <3>
#> [7] CCC                                                                      <3>
#> [8] CTG                                                                      <3>
#> 

# Extracting second, fourth, third and second letters:
bite(sq, c(2,4,3,2))

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] CTTC                                                                     <4>
#> [2] TTTT                                                                     <4>
#> [3] CTTC                                                                     <4>
#> [4] CTTC                                                                     <4>
#> [5] CGTC                                                                     <4>
#> [6] TATT                                                                     <4>
#> [7] CTCC                                                                     <4>
#> [8] TAGT                                                                     <4>
#> 

# Extracting second to fifth letter - NA introduced:
bite(sq, 2:5)

#> Warning: some sequences are subsetted with index bigger than length - NA introduced

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] CTTT                                                                     <4>
#> [2] TTTG                                                                     <4>
#> [3] CTTT                                                                     <4>
#> [4] CTTT                                                                     <4>
#> [5] CTGC                                                                     <4>
#> [6] TTAG                                                                     <4>
#> [7] CCT!                                                                     <4>
#> [8] TGAA                                                                     <4>
#> 

# Extracting all from first to twentieth - NA introduced:
bite(sq, 1:20)

#> Warning: some sequences are subsetted with index bigger than length - NA introduced

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TCTTTGGTTATCTAGCTGTA                                                    <20>
#> [2] CTTTGGTTATCTAGCTGTAT                                                    <20>
#> [3] TCTTTGGTTATCTAGCTGTA                                                    <20>
#> [4] TCTTTGGTTATCTAGCTGTA                                                    <20>
#> [5] ACTGCTG!!!!!!!!!!!!!                                                    <20>
#> [6] CTTAGA!!!!!!!!!!!!!!                                                    <20>
#> [7] CCCT!!!!!!!!!!!!!!!!                                                    <20>
#> [8] CTGAATGT!!!!!!!!!!!!                                                    <20>
#> 

# Extracting all excluding first letter of sequence:
bite(sq, -1)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] CTTTGGTTATCTAGCTGTATGA                                                  <22>
#> [2] TTTGGTTATCTAGCTGTATGA                                                   <21>
#> [3] CTTTGGTTATCTAGCTGTATG                                                   <21>
#> [4] CTTTGGTTATCTAGCTGTATGAA                                                 <23>
#> [5] CTGCTG                                                                   <6>
#> [6] TTAGA                                                                    <5>
#> [7] CCT                                                                      <3>
#> [8] TGAATGT                                                                  <7>
#> 

# Extracting all excluding second and sixth letter of sequence:
bite(sq, c(-2, -6))

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TTTTGTTATCTAGCTGTATGA                                                   <21>
#> [2] CTTGTTATCTAGCTGTATGA                                                    <20>
#> [3] TTTTGTTATCTAGCTGTATG                                                    <20>
#> [4] TTTTGTTATCTAGCTGTATGAA                                                  <22>
#> [5] ATGCG                                                                    <5>
#> [6] CTAG                                                                     <4>
#> [7] CCT                                                                      <3>
#> [8] CGAAGT                                                                   <6>
#>