Create complement sequence from nucsq object

Creates the complementary sequence from a given RNA or DNA nucleotide sequence. The function differentiates between RNA and DNA sequences.

complement(nucsq, is_dna = NULL)

complement_dna(nucsq)

complement_rna(nucsq)

Arguments

nucsq
a sq object of type nuc.

is_dna
a logical indicating if entered sequence is DNA or RNA. If is_dna is TRUE, sequences cannot contain 'U' (uracil); if is FALSE, sequences cannot contain 'T' (thymine). If NULL (default) the sequence type is not specified and is guessed (see details below).

Value

sq object of the same type as input nucsq (nuc) but built of complementary nucleotides to entered sequence.

Details

This functions allows to obtain a complement sequence, which is created by matching elements (nucleotides) with complementarity to input nucsq object. In case of complement: If is_dna = TRUE entered sequence is DNA. If is_dna = FALSE entered sequence is RNA. By default, the sequence type is not specified and the function tries to guess, which type of sequence was entered. If a sequence contains 'U' without 'T' the type will be set to RNA. If a sequence contains 'T' (thymine) without 'U' (uracil), the type is set to DNA. An error is displayed if both 'T' and 'U' are present in the sequence or if the sequence contains only 'A' (adenine). If the sequence does not contain 'T' or 'U' or the logical specification is wrong (i. e. if the sequence contains 'U' and the logical specification is set to DNA), an error will also be returned.

Functions complement_dna and complement_rna are wrappers around complement with specified is_dna parameter - accordingly TRUE or FALSE.

See also

Examples

# Creating objects nuc_dna_sequence (with DNA sequences) and nuc_rna_sequence 
# (with RNA sequences) to work on:

nuc_dna_sequence <- construct_sq(c("ACTGCTG", "CTTAGA",
                                   "CCCT", "CTGAATGT"), type = "nuc")
nuc_rna_sequence <- construct_sq(c("ACUGCUG", "CUUAGA",
                                   "CCCU", "CUGAAUGU"), type = "nuc")

# In the following the complement function is used to make a PCR (Polymerase Chain Reaction)
# primers. Every sequence will be rewritten to its complementary equivalent as 
# following example: AAATTTGGG to TTTAAACCC.

# creating complementary sequences with a specified sequence type (and using wrappers):
complement(nuc_dna_sequence, is_dna = TRUE)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TGACGAC                                                                  <7>
#> [2] GAATCT                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACTTACA                                                                 <8>
#> 

complement_dna(nuc_dna_sequence)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TGACGAC                                                                  <7>
#> [2] GAATCT                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACTTACA                                                                 <8>
#> 

complement(nuc_rna_sequence, is_dna = FALSE)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] UGACGAC                                                                  <7>
#> [2] GAAUCU                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACUUACA                                                                 <8>
#> 

complement_rna(nuc_rna_sequence)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] UGACGAC                                                                  <7>
#> [2] GAAUCU                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACUUACA                                                                 <8>
#> 

# creating complementary sequences without defined sequence type:
complement(nuc_dna_sequence)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] TGACGAC                                                                  <7>
#> [2] GAATCT                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACTTACA                                                                 <8>
#> 

complement(nuc_rna_sequence)

#> nuc (nucleotides), cln (cleaned) sequences list:
#> [1] UGACGAC                                                                  <7>
#> [2] GAAUCU                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACUUACA                                                                 <8>
#> 

# Each sequence from nuc_dna_sequence and nuc_rna_sequence object have now an  
# own complementary equivalent, which can be helpful during constructing PCR primers.