Create complement sequence from dnasq or rnasq object

Creates the complementary sequence from a given RNA or DNA nucleotide sequence. The function differentiates between RNA and DNA sequences.

complement(sq)

complement_dna(sq)

complement_rna(sq)

Arguments

sq
a sq object of type dna or rna.

Value

sq object of the same type as input dnasq (dna) or rnasq (rna) but built of complementary nucleotides to entered sequence.

Details

This function allows to get complement sequence, which is created by matching elements (nucleotides) with complementary to input dnasq or rnasq object. Whether 'U' (uracil) or 'T' (thymine) is used depends on the class of the sq object.

Functions complement_dna and complement_rna are provided as a safe way of limiting classes complement function is used on.

See also

Examples

# Creating objects dna_sequence (with DNA sequences) and rna_sequence 
# (with RNA sequences) to work on:

dna_sequence <- construct_sq(c("ACTGCTG", "CTTAGA", "CCCT", "CTGAATGT"),
                             type = "dna")
rna_sequence <- construct_sq(c("ACUGCUG", "CUUAGA", "CCCU", "CUGAAUGU"),
                             type = "rna")

# In the following the complement function is used to make a PCR (Polymerase Chain Reaction)
# primers. Every sequence will be rewritten to its complementary equivalent as 
# following example: AAATTTGGG to TTTAAACCC.
 
# creating complementary sequences with the basic function and using wrappers:
complement(dna_sequence)

#> dna (DNA), cln (cleaned) sequences list:

#> [1] TGACGAC                                                                  <7>
#> [2] GAATCT                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACTTACA                                                                 <8>

complement_dna(dna_sequence)

#> dna (DNA), cln (cleaned) sequences list:

#> [1] TGACGAC                                                                  <7>
#> [2] GAATCT                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACTTACA                                                                 <8>

complement(rna_sequence)

#> rna (RNA), cln (cleaned) sequences list:

#> [1] UGACGAC                                                                  <7>
#> [2] GAAUCU                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACUUACA                                                                 <8>

complement_rna(rna_sequence)

#> rna (RNA), cln (cleaned) sequences list:

#> [1] UGACGAC                                                                  <7>
#> [2] GAAUCU                                                                   <6>
#> [3] GGGA                                                                     <4>
#> [4] GACUUACA                                                                 <8>

# Each sequence from dna_sequence and rna_sequence object have now an own
# complementary equivalent, which can be helpful during constructing PCR primers.