Construct sq object from character vector

This function allows the user to construct objects of class sq from a character vector.

construct_sq(sq, type = NULL, is_clean = NULL, non_standard = NULL)

construct_sq_ami(sq, is_clean = NULL)

construct_sq_dna(sq, is_clean = NULL)

construct_sq_rna(sq, is_clean = NULL)

Arguments

sq
a character vector.

type
a character string indicating type of sq object that is going to be constructed; supported values are "ami" for amino acid sequences, "dna" for DNA sequences, "rna" for RNA sequences, "unt" for and NULL for type guessing (see details)

is_clean
a logical value indicating if sequences are clean. or in other words - they don't contain ambiguous values; supported values are TRUE for clean sequences, FALSE for unclean sequences and NULL for auto detecting (see details).

non_standard
a character vector indicating non-standard letters contained in sequences. If NULL, sequences will not be searched for non-standard letters of length more than one. Each element of this parameter should be at least two characters long.

Value

object of class sq with appropriate type (one of: ami, dna, rna, unt, atp).

Details

Function construct_sq covers all possibilities of standard and non-standard types and alphabets. You can check what 'type' and 'alphabet' exactly are in sq class documentation. Below there is a guide how function operates and how the program behaves according to the given arguments and the letters in the sequences.

Functions construct_sq_ami, construct_sq_dna and construct_sq_rna are wrappers around construct_sq with specified type parameter - accordingly "ami", "dna" or "rna". You can also pass "is_clean" parameter to those functions, but you cannot pass "non_standard".

sq parameter should be a character vector. Each element of this vector is a biological sequence. If this parameter has length 0, object of class sq with 0 sequences will be created (if not specified, it will have dna cln type, which is a result of rules written below). If it contains sequences of length 0, NULL sequences will be introduced (see NULL (empty) sequences section in sq class).

Important note: in all below cases word 'letter' stands for an element of an alphabet. Letter might consist of more than one character, for example "Ala" might be a single letter. However, if you want to construct or read sequences with multi-character letters, one has to specify non_standard parameter. Details of letters, alphabet and types can be found in sq class documentation.

Simple guide to construct

In most cases, just the sq parameter needs to be specified - type of sequences will be guessed accordingly to rules described below. You need to pay attention, however, because for short sequences type may be guessed incorrectly - in this case you should specify type and/or is_clean.

If your sequences contain non-standard letters, where each non-standard letter is one character long, you also don't need to specify any parameter. Optionally, you can explicitly do it by setting type to "unt".

If you want to construct sequences with multicharacter letters, you have to specify non_standard parameter, where you have to provide all non-standard letters longer than one character.

In fast mode you have to specify both type and is_clean parameters. You cannot specify non_standard parameter in this mode. All letters outside specified alphabet will be red as NA values.

Detailed guide to construct

Below all possibilities that can occur during the construction of a sq object are listed.

In normal mode (no fast mode):

If you don't specify any other parameter than sq, function will try to guess sequence type and if it is clean (it will check in exactly this order):

If it contains only ACGT- letters, either lowercase or uppercase, type will be set to dna cln.

If it contains only ACGU- letters, either lowercase or uppercase, type will be set to dna cln.

If it contains any letters from 1. and 2. and additionally letters DEFHIKLMNPQRSVWY*, either lowercase or uppercase, type will be set to ami cln.

If it contains any letters from 1. and additionally letters WSMKRYBDHVN, either lowercase or uppercase, and does not contain any other letter, type will be set to dna without cln subtype.

If it contains any letters from 2. and additionally letters WSMKRYBDHVN, either lowercase or uppercase, and does not contain any other letter, type will be set to rna without cln subtype.

If it contains any letters from previous points and additionally letters JOUXZ, type will be set to ami without cln subtype.

If it contains any letters that exceed all groups mentioned above, type will be set to "unt".

If you specify type parameter as "ami", "dna" or "rna" (and do not specify neither is_clean nor non_standard) type will be checked during construction:

If all letters in sequences fit the clean alphabet of given type, type will be set to given type with cln subtype.

If all letters in sequences fit the unclean alphabet of given type, type will be set to given type without cln subtype.

If at least one sequence contains at least one letter that is not an element of an unclean alphabet of the provided type, an error will be thrown.

If you specify both type and is_clean, function checks if letters in sequences matches exactly specified alphabet (with capitalisation accuracy). If they do, type will be set to it. Otherwise, an error will be thrown.

If you specify type as "unt" and neither is_clean nor non_standard, type will be set to unt. Letters won't be converted to uppercase, alphabet will consist of all letters found in sequences.

If you do not specify neither type nor is_clean and specify non_standard parameter, which should be character vector where each element is at least two characters long, all strings as specified will be detected in sequences and treated as letters in constructed atp sq.

All other combinations of parameters are incorrect.

In fast mode you have to specify type (it has to have one of "ami", "dna" or "rna" values) and is_clean (TRUE or FALSE). You cannot specify non_standard. All letters that aren't elements of destination alphabet (with a letter size accuracy) will be treated as NA values.

Handling atp and unt sequences and NA values

You can convert letters into another using substitute_letters and then you can use typify function to set of sq to ami, dna or rna. If your sequences contain NA values, use remove_na

See also

Examples

# saving option:
previous_option <- getOption("tidysq_g_fast_mode")

#### constructing sq in normal mode:
## setting an option:
options(tidysq_g_fast_mode = FALSE)

## constructing sq without specifying type
# dna cln sq
construct_sq(c("ATGC", "TCGTTA", "TT--AG"))

#> dna (DNA), cln (cleaned) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

# rna cln sq
construct_sq(c("CUUAC", "UACCGGC", "GCA-ACGU"))

#> rna (RNA), cln (cleaned) sequences list:
#> [1] CUUAC                                                                    <5>
#> [2] UACCGGC                                                                  <7>
#> [3] GCA-ACGU                                                                 <8>

# ami cln sq
construct_sq(c("YQQPAVVM", "PQCFL"))

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] YQQPAVVM                                                                 <8>
#> [2] PQCFL                                                                    <5>

# ami cln sq can contain * - letter meaning end of translation:
construct_sq(c("MMDF*", "SYIHR*", "MGG*"))

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] MMDF*                                                                    <5>
#> [2] SYIHR*                                                                   <6>
#> [3] MGG*                                                                     <4>

# dna sq
construct_sq(c("TMVCCDA", "BASDT-CNN"))

#> dna (DNA) sequences list:
#> [1] TMVCCDA                                                                  <7>
#> [2] BASDT-CNN                                                                <9>

# rna sq
construct_sq(c("WHDHKYN", "GCYVCYU"))

#> rna (RNA) sequences list:
#> [1] WHDHKYN                                                                  <7>
#> [2] GCYVCYU                                                                  <7>

# ami sq
construct_sq(c("XYOQWWKCNJLO"))

#> ami (amino acids) sequences list:
#> [1] XYOQWWKCNJLO                                                            <12>

# unt sq - let's assume that one wants to mark some special element in sequence with %
construct_sq(c("%%YAPLAA", "PLAA"))

#> unt (unspecified type) sequences list:
#> [1] %%YAPLAA                                                                 <8>
#> [2] PLAA                                                                     <4>

## constructing sq with type 
# all above examples will result in an identical if specified type as guessed
construct_sq(c("ATGC", "TCGTTA", "TT--AG"), "dna")

#> dna (DNA), cln (cleaned) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

construct_sq(c("CUUAC", "UACCGGC", "GCA-ACGU"), "rna")

#> rna (RNA), cln (cleaned) sequences list:
#> [1] CUUAC                                                                    <5>
#> [2] UACCGGC                                                                  <7>
#> [3] GCA-ACGU                                                                 <8>

construct_sq(c("YQQPAVVM", "PQCFL"), "ami")

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] YQQPAVVM                                                                 <8>
#> [2] PQCFL                                                                    <5>

construct_sq(c("MMDF*", "SYIHR*", "MGG*"), "ami")

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] MMDF*                                                                    <5>
#> [2] SYIHR*                                                                   <6>
#> [3] MGG*                                                                     <4>

construct_sq(c("TMVCCDA", "BASDT-CNN"), "dna")

#> dna (DNA) sequences list:
#> [1] TMVCCDA                                                                  <7>
#> [2] BASDT-CNN                                                                <9>

construct_sq(c("WHDHKYN", "GCYVCYU"), "rna")

#> rna (RNA) sequences list:
#> [1] WHDHKYN                                                                  <7>
#> [2] GCYVCYU                                                                  <7>

construct_sq(c("XYOQWWKCNJLO"), "ami")

#> ami (amino acids) sequences list:
#> [1] XYOQWWKCNJLO                                                            <12>

construct_sq(c("%%YAPLAA", "PLAA"), "unt")

#> unt (unspecified type) sequences list:
#> [1] %%YAPLAA                                                                 <8>
#> [2] PLAA                                                                     <4>

# you can also use wrappers instead of parameters
construct_sq_dna(c("ATGC", "TCGTTA", "TT--AG"))

#> dna (DNA), cln (cleaned) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

construct_sq_rna(c("CUUAC", "UACCGGC", "GCA-ACGU"))

#> rna (RNA), cln (cleaned) sequences list:
#> [1] CUUAC                                                                    <5>
#> [2] UACCGGC                                                                  <7>
#> [3] GCA-ACGU                                                                 <8>

construct_sq_ami(c("YQQPAVVM", "PQCFL"))

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] YQQPAVVM                                                                 <8>
#> [2] PQCFL                                                                    <5>

construct_sq_ami(c("MMDF*", "SYIHR*", "MGG*"))

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] MMDF*                                                                    <5>
#> [2] SYIHR*                                                                   <6>
#> [3] MGG*                                                                     <4>

construct_sq_dna(c("TMVCCDA", "BASDT-CNN"))

#> dna (DNA) sequences list:
#> [1] TMVCCDA                                                                  <7>
#> [2] BASDT-CNN                                                                <9>

construct_sq_rna(c("WHDHKYN", "GCYVCYU"))

#> rna (RNA) sequences list:
#> [1] WHDHKYN                                                                  <7>
#> [2] GCYVCYU                                                                  <7>

construct_sq_ami(c("XYOQWWKCNJLO"))

#> ami (amino acids) sequences list:
#> [1] XYOQWWKCNJLO                                                            <12>

# One can force type other than guessed (if letters fit in the destination alphabet)
construct_sq(c("ATGC", "TCGTTA", "TT--AG"), "dna", is_clean = FALSE)

#> dna (DNA) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

construct_sq(c("ATGC", "TCGTTA", "TT--AG"), "ami")

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

construct_sq(c("ATGC", "TCGTTA", "TT--AG"), "ami", is_clean = FALSE)

#> ami (amino acids) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

construct_sq(c("ATGC", "TCGTTA", "TT--AG"), "unt")

#> unt (unspecified type) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

## constructing with non_standard specified
# in sequences below "mA" denotes methyled alanine - two characters are treated as single letter
construct_sq(c("LmAQYmASSR", "LmASMKLKFmAmA"), non_standard = "mA")

#> atp (atypical alphabet) sequences list:
#> [1] L mA Q Y mA S S R                                                        <8>
#> [2] L mA S M K L K F mA mA                                                  <10>

# reading sequences with three-letter names:
construct_sq(c("ProProGlyAlaMetAlaCys"), non_standard = c("Pro", "Gly", "Ala", "Met", "Cys"))

#> atp (atypical alphabet) sequences list:
#> [1] Pro Pro Gly Ala Met Ala Cys                                              <7>

#### constructing in fast mode:
## setting fast mode on
options(tidysq_g_fast_mode = TRUE)

# you cannot construct without specifying type
if (FALSE) {
construct_sq("CTGA")
}
construct_sq("CTGA", "dna", TRUE)

#> dna (DNA), cln (cleaned) sequences list:
#> [1] CTGA                                                                     <4>

# you cannot construct with specifying non_standard
if (FALSE) {
construct_sq("mAPQ", non_standard = "mA")
}

# letters other than in specified alphabet will be treated as NA:
construct_sq("NCTGCNA", "dna", TRUE)

#> dna (DNA), cln (cleaned) sequences list:
#> [1] !CTGC!A                                                                  <7>

#### Other examples:
## setting fast mode off again:
options(tidysq_g_fast_mode = FALSE)

# lowercase letters are converted to uppercase if detected type is
# "ami", "dna" or "rna"
construct_sq(c("aTGc", "tcgTTA", "tt--AG"))

#> dna (DNA), cln (cleaned) sequences list:
#> [1] ATGC                                                                     <4>
#> [2] TCGTTA                                                                   <6>
#> [3] TT--AG                                                                   <6>

construct_sq(c("XYOqwwKCNJLo"))

#> ami (amino acids) sequences list:
#> [1] XYOQWWKCNJLO                                                            <12>

# but not for "unt"
construct_sq(c("aAAaAA"), type = "unt")

#> unt (unspecified type) sequences list:
#> [1] aAAaAA                                                                   <6>

# you can construct sq with length 0
construct_sq(character(0))

#> dna (DNA), cln (cleaned) sequences list of length 0

# and sq with empty sequences
construct_sq(c("AGTGGC", "", "CATGA", ""))

#> dna (DNA), cln (cleaned) sequences list:
#> [1] AGTGGC                                                                   <6>
#> [2] <NULL>                                                                   <0>
#> [3] CATGA                                                                    <5>
#> [4] <NULL>                                                                   <0>

## reseting an option
options(tidysq_g_fast_mode = previous_option)