Encode sequences as numeric values

Encode sequences, stored in sq object, with numeric values defined by user in named vector.

encode(sq, encoding)

Arguments

sq
a sq object.

encoding
encoding a named numeric vector, that consists of values assigned to nucleic or amino acid letters. Letters without assigned encoding will be shown as NA. Names of vector should be unique and should be elements of alphabet of sq object.

Value

sq object of typeenc.

Details

Each position in a sequence is replaced by a numeric value assigned to that letter.

Sometimes for research purposes one wants to replace letters by values, described by various properties of nucleic or amino acids. It can be a sparse encoding, hydrophobicity scale, heat capacities, entropies, chemical shift index, probability matrix (BLOSUM, PAM), sequence profile or reduced alphabet.

The newly constructed sequence will have a type enc (see details on types in sq), which represents encoded sequences.

The named vector (ex. c(G = 1, K = 2, P = 2)) should have all letters assigned, otherwise unassigned letters will be shown as NA. If any letter that appears in an alphabet appears in at least one of sequences, user will be informed about it. Default action is a warning printed in the console, but it can be changed via setting "tidysq_a_no_given_enc" (see details at tidysq-options).

In fact the only thing that is replaced is an alphabet - letters are substituted with values assigned to them. The internal structure of the object remains unchanged.

If one wants to access numeric values of encoded sequences, they may use as.matrix or encsq_to_list.

See also

Examples

# Create sq object with sequences containing letters from 
# standard alphabet to work on:

sq_dna <- construct_sq(c("TATGAATTAGCTGTCTTTGCTGCTTTGGTTATCTATGA",
                         "CTTTGGTTATCTAGCTGTATGA", "TATCTAGCTGTATG",
                         "CTGCTG", "CTTAGA", "CCCT", "CTGAATGT"),
                       type = "dna")

sq_ami <- construct_sq(c("NYMITGGREEYERTVIYRAIALNAANYTWIL",
                        "TIAALGNIIYRAIE", "NYERTGHLI",
                        "MAYNNNIALN", "MN", "NAAAT"),
                        type = "ami")

# Create an object, called sq, with sequences containing letters from 
# standard and extended alphabet to work on:

sq_rna_ex <- construct_sq(c("VAHCHAGDUGBBVG", "CUGCVB", "DUUBGA", "CCCU",
                            "CUHAABBU"), type = "rna")

sq_ami_ex <- construct_sq(c("MAYUOUONNNIALN", "UUMXBZONO", "NAAGAT"), type = "ami")

# Create encoding for a standard alphabet 

enc_dna <- c(A = 1, C = 2, G = 2, T = 2)
enc_rna <- c(A = 7, C = 5, G = 3, U = 2)
enc_ami <- c(A = 5, C = 5, D = 6, E = 6, F = 4,
             G = 1, H = 5, I = 3, K = 2, L = 3,
             M = 5, N = 6, P = 2, Q = 6, R = 2,
             S = 6, T = 6, V = 3, W = 4, Y = 4)

# Create encoding for an extended alphabet

enc_dna_ex <- c(A = 1, C = 1, G = 2, T = 3,
                B = 5, D = 8, H = 13, V = 21)
enc_rna_ex <- c(A = 1, C = 1, G = 2, U = 3,
                B = 5, D = 8, H = 13, V = 21)
enc_ami_ex <- c(A = 5, C = 5, D = 6, E = 6, F = 4,
                G = 1, H = 5, I = 3, K = 2, L = 3,
                M = 5, N = 6, P = 2, Q = 6, R = 2,
                S = 6, T = 6, V = 3, W = 4, Y = 4,
                U = -0.1, O = 0.753, X = -53.95, B = 7.77, Z = 0)

# Encode with simplified standard alphabet sequences without 
# non-standard letters

encode(sq_dna, enc_dna)

#> enc (encoded values) sequences list:
#> [1]  2  1  2  2  1  1  2  2  1  2  2  2  2  2  2  2  2  2  2  2  2  2  2... <38>
#> [2]  2  2  2  2  2  2  2  2  1  2  2  2  1  2  2  2  2  2  1  2  2  1       <22>
#> [3]  2  1  2  2  2  1  2  2  2  2  2  1  2  2                               <14>
#> [4]  2  2  2  2  2  2                                                        <6>
#> [5]  2  2  2  1  2  1                                                        <6>
#> [6]  2  2  2  2                                                              <4>
#> [7]  2  2  2  1  1  2  2  2                                                  <8>

encode(sq_ami, enc_ami)

#> enc (encoded values) sequences list:
#> [1]  6  4  5  3  6  1  1  2  6  6  4  6  2  6  3  3  4  2  5  3  5  3  6... <31>
#> [2]  6  3  5  5  3  1  6  3  3  4  2  5  3  6                               <14>
#> [3]  6  4  6  2  6  1  5  3  3                                               <9>
#> [4]  5  5  4  6  6  6  3  5  3  6                                           <10>
#> [5]  5  6                                                                    <2>
#> [6]  6  5  5  5  6                                                           <5>

# Encode with simplified extended alphabet sequences without 
# non-standard letters

encode(sq_dna, enc_dna_ex)

#> enc (encoded values) sequences list:
#> [1]  3  1  3  2  1  1  3  3  1  2  1  3  2  3  1  3  3  3  2  1  3  2  1... <38>
#> [2]  1  3  3  3  2  2  3  3  1  3  1  3  1  2  1  3  2  3  1  3  2  1       <22>
#> [3]  3  1  3  1  3  1  2  1  3  2  3  1  3  2                               <14>
#> [4]  1  3  2  1  3  2                                                        <6>
#> [5]  1  3  3  1  2  1                                                        <6>
#> [6]  1  1  1  3                                                              <4>
#> [7]  1  3  2  1  1  3  2  3                                                  <8>

encode(sq_ami, enc_ami_ex)

#> enc (encoded values) sequences list:
#> [1]  6  4  5  3  6  1  1  2  6  6  4  6  2  6  3  3  4  2  5  3  5  3  6... <31>
#> [2]  6  3  5  5  3  1  6  3  3  4  2  5  3  6                               <14>
#> [3]  6  4  6  2  6  1  5  3  3                                               <9>
#> [4]  5  5  4  6  6  6  3  5  3  6                                           <10>
#> [5]  5  6                                                                    <2>
#> [6]  6  5  5  5  6                                                           <5>

# Encode with simplified standard alphabet sequences with 
# non-standard letters

encode(sq_rna_ex, enc_rna)

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

#> enc (encoded values) sequences list:
#> [1] NA  7 NA  5 NA  7  3 NA  2  3 NA NA NA  3                               <14>
#> [2]  5  2  3  5 NA NA                                                        <6>
#> [3] NA  2  2 NA  3  7                                                        <6>
#> [4]  5  5  5  2                                                              <4>
#> [5]  5  2 NA  7  7 NA NA  2                                                  <8>

encode(sq_ami_ex, enc_ami)

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

#> enc (encoded values) sequences list:
#> [1]  5  5  4 NA NA NA NA  6  6  6  3  5  3  6                               <14>
#> [2] NA NA  5 NA NA NA NA  6 NA                                               <9>
#> [3]  6  5  5  1  5  6                                                        <6>

# Encode with simplified extended alphabet sequences with 
# non-standard letters

encode(sq_rna_ex, enc_rna_ex)

#> enc (encoded values) sequences list:
#> [1] 21  1 13  1 13  1  2  8  3  2  5  5 21  2                               <14>
#> [2]  1  3  2  1 21  5                                                        <6>
#> [3]  8  3  3  5  2  1                                                        <6>
#> [4]  1  1  1  3                                                              <4>
#> [5]  1  3 13  1  1  5  5  3                                                  <8>

encode(sq_ami_ex, enc_ami_ex)

#> enc (encoded values) sequences list:
#> [1]   5.00   5.00   4.00  -0.10   0.75  -0.10   0.75   6.00   6.00...       <14>
#> [2]  -0.10  -0.10   5.00 -53.95   7.77   0.00   0.75   6.00   0.75           <9>
#> [3]   6.00   5.00   5.00   1.00   5.00   6.00                                <6>

# Encode without assigning all letters

encode(sq_dna, c(A = 1, G = 0.02))

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

#> enc (encoded values) sequences list:
#> [1]   NA 1.00   NA 0.02 1.00 1.00   NA   NA 1.00 0.02   NA   NA 0.02...     <38>
#> [2]   NA   NA   NA   NA 0.02 0.02   NA   NA 1.00   NA   NA   NA 1.00...     <22>
#> [3]   NA 1.00   NA   NA   NA 1.00 0.02   NA   NA 0.02   NA 1.00   NA 0.02   <14>
#> [4]   NA   NA 0.02   NA   NA 0.02                                            <6>
#> [5]   NA   NA   NA 1.00 0.02 1.00                                            <6>
#> [6]   NA   NA   NA   NA                                                      <4>
#> [7]   NA   NA 0.02 1.00 1.00   NA 0.02   NA                                  <8>

encode(sq_ami, c(A = 5, H = 5, I = 0.3, K = -2, L = -3.1, M = 5, N = 6))

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

#> enc (encoded values) sequences list:
#> [1]  6.0   NA  5.0  0.3   NA   NA   NA   NA   NA   NA   NA   NA   NA...     <31>
#> [2]   NA  0.3  5.0  5.0 -3.1   NA  6.0  0.3  0.3   NA   NA  5.0  0.3   NA   <14>
#> [3]  6.0   NA   NA   NA   NA   NA  5.0 -3.1  0.3                             <9>
#> [4]  5.0  5.0   NA  6.0  6.0  6.0  0.3  5.0 -3.1  6.0                       <10>
#> [5]  5.0  6.0                                                                <2>
#> [6]  6.0  5.0  5.0  5.0   NA                                                 <5>

# Import encoding from \code{AAindex database}

data("AAindex_norm")
enc_aa <- AAindex_norm[20, ]

encode(sq_ami, enc_aa)

#> enc (encoded values) sequences list:
#> [1] 0.087 0.688 0.526 0.634 0.565 0.336 0.336 0.486 0.186 0.186 0.688...    <31>
#> [2] 0.565 0.634 0.264 0.264 0.601 0.336 0.087 0.634 0.634 0.688 0.486...    <14>
#> [3] 0.087 0.688 0.186 0.486 0.565 0.336 0.000 0.601 0.634                    <9>
#> [4] 0.526 0.264 0.688 0.087 0.087 0.087 0.634 0.264 0.601 0.087             <10>
#> [5] 0.526 0.087                                                              <2>
#> [6] 0.087 0.264 0.264 0.264 0.565                                            <5>