Transform encoded sequence to a list

Transform encoded by encode sequence, stored in a sq object, to a list.

encsq_to_list(encsq)

Arguments

encsq
a sq object with enc type, which was encoded using encode function.

In a new object you can check what value is assigned to each letter.

Value

A list with named numeric vectors. Each vector represents values assigned to according sequence with encode function. Vectors are named and their names are letters of original sequences.

Details

Function is used to transform an sq object with enc type to a list, which can be more useful in some cases.

See also

Examples

# Create sq object with sequences containing letters from 
# standard alphabet and extended alphabet to work on:

sq_dna <- construct_sq(c("TATGAATTAGCTGTCTTTGCTGCTTTGGTTATCTATGA", 
                         "CTTTGGTTATCTAGCTGTATGA", "TATCTAGCTGTATG", 
                         "CTGCTG", "CTTAGA", "CCCT", "CTGAATGT"), 
                       type = "dna")

sq_ami <- construct_sq(c("NYMITGGREEYERTVIYRAIALNAANYTWIL", 
                         "TIAALGNIIYRAIE", "NYERTGHLI", "MAYNNNIALN", 
                         "MN", "NAAAT"), type = "ami")

sq_rna_ex <- construct_sq(c("VAHCHAGDUGBBVG", "CUGCVB", "DUUBGA", "CCCU", 
                            "CUHAABBU"), type = "rna")

sq_ami_ex <- construct_sq(c("MAYUOUONNNIALN", "UUMXBZONO", 
                            "NAAGAT"))

# Create encoding for standard, extended alphabet and import from 
# other sources:

enc_dna <- c(A = 1, C = 2, G = 2, T = 2)
enc_rna <- c(A = 7, C = 5, G = 3, U = 2)
enc_ami <- c(A = 5, C = 5, D = 6, E = 6, F = 4, 
             G = 1, H = 5, I = 3, K = 2, L = 3, 
             M = 5, N = 6, P = 2, Q = 6, R = 2, 
             S = 6, T = 6, V = 3, W = 4, Y = 4)
             
enc_dna_ex <- c(A = 1, C = 1, G = 2, T = 3,
                B = 5, D = 8, H = 13, V = 21)
enc_rna_ex <- c(A = 1, C = 1, G = 2, U = 3,
                B = 5, D = 8, H = 13, V = 21)
enc_ami_ex <- c(A = 5, C = 5, D = 6, E = 6, F = 4, 
                G = 1, H = 5, I = 3, K = 2, L = 3, 
                M = 5, N = 6, P = 2, Q = 6, R = 2, 
                S = 6, T = 6, V = 3, W = 4, Y = 4, 
                U = -0.1, O = 0.753, X = -53.95, 
                B = 7.77, Z = 0)
                
data("AAindex")
enc_aa <- AAindex[20, ]

# Encode sequences and assign it to a variable:

e1 <- encode(sq_dna, enc_dna)
e2 <- encode(sq_ami, enc_ami)
e3 <- encode(sq_dna, enc_rna_ex)

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

e4 <- encode(sq_ami, enc_ami_ex)
e5 <- encode(sq_rna_ex, enc_dna)

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

e6 <- encode(sq_ami_ex, enc_ami)

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

e7 <- encode(sq_rna_ex, enc_rna_ex)
e8 <- encode(sq_ami_ex, enc_ami_ex)
e9 <- encode(sq_dna, c(A = 1, G = 0.02))

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

e10 <- encode(sq_ami, c(A = 5, H = 5, I = 0.3, K = -2, 
                        L = -3.1, M = 5, N = 6))

#> Warning: there are letters in the alphabet of 'sq' that appear in sequences, but were not given in 'encoding' - assuming NA

e_aa <- encode(sq_ami, enc_aa)

# Transform \code{\link{encode}} result to a list:

## Sequence with a standard alphabet, encoding with a standard alphabet
encsq_to_list(e1)

#> [[1]]
#> T A T G A A T T A G C T G T C T T T G C T G C T T T G G T T A T C T A T G A 
#> 2 1 2 2 1 1 2 2 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 2 2 2 1 2 2 1 
#> 
#> [[2]]
#> C T T T G G T T A T C T A G C T G T A T G A 
#> 2 2 2 2 2 2 2 2 1 2 2 2 1 2 2 2 2 2 1 2 2 1 
#> 
#> [[3]]
#> T A T C T A G C T G T A T G 
#> 2 1 2 2 2 1 2 2 2 2 2 1 2 2 
#> 
#> [[4]]
#> C T G C T G 
#> 2 2 2 2 2 2 
#> 
#> [[5]]
#> C T T A G A 
#> 2 2 2 1 2 1 
#> 
#> [[6]]
#> C C C T 
#> 2 2 2 2 
#> 
#> [[7]]
#> C T G A A T G T 
#> 2 2 2 1 1 2 2 2 
#> 

encsq_to_list(e2)

#> [[1]]
#> N Y M I T G G R E E Y E R T V I Y R A I A L N A A N Y T W I L 
#> 6 4 5 3 6 1 1 2 6 6 4 6 2 6 3 3 4 2 5 3 5 3 6 5 5 6 4 6 4 3 3 
#> 
#> [[2]]
#> T I A A L G N I I Y R A I E 
#> 6 3 5 5 3 1 6 3 3 4 2 5 3 6 
#> 
#> [[3]]
#> N Y E R T G H L I 
#> 6 4 6 2 6 1 5 3 3 
#> 
#> [[4]]
#> M A Y N N N I A L N 
#> 5 5 4 6 6 6 3 5 3 6 
#> 
#> [[5]]
#> M N 
#> 5 6 
#> 
#> [[6]]
#> N A A A T 
#> 6 5 5 5 6 
#> 

## Sequence with a standard alphabet, encoding with an extended alphabet
encsq_to_list(e3)

#> [[1]]
#>  T  A  T  G  A  A  T  T  A  G  C  T  G  T  C  T  T  T  G  C  T  G  C  T  T  T 
#> NA  1 NA  2  1  1 NA NA  1  2  1 NA  2 NA  1 NA NA NA  2  1 NA  2  1 NA NA NA 
#>  G  G  T  T  A  T  C  T  A  T  G  A 
#>  2  2 NA NA  1 NA  1 NA  1 NA  2  1 
#> 
#> [[2]]
#>  C  T  T  T  G  G  T  T  A  T  C  T  A  G  C  T  G  T  A  T  G  A 
#>  1 NA NA NA  2  2 NA NA  1 NA  1 NA  1  2  1 NA  2 NA  1 NA  2  1 
#> 
#> [[3]]
#>  T  A  T  C  T  A  G  C  T  G  T  A  T  G 
#> NA  1 NA  1 NA  1  2  1 NA  2 NA  1 NA  2 
#> 
#> [[4]]
#>  C  T  G  C  T  G 
#>  1 NA  2  1 NA  2 
#> 
#> [[5]]
#>  C  T  T  A  G  A 
#>  1 NA NA  1  2  1 
#> 
#> [[6]]
#>  C  C  C  T 
#>  1  1  1 NA 
#> 
#> [[7]]
#>  C  T  G  A  A  T  G  T 
#>  1 NA  2  1  1 NA  2 NA 
#> 

encsq_to_list(e4)

#> [[1]]
#> N Y M I T G G R E E Y E R T V I Y R A I A L N A A N Y T W I L 
#> 6 4 5 3 6 1 1 2 6 6 4 6 2 6 3 3 4 2 5 3 5 3 6 5 5 6 4 6 4 3 3 
#> 
#> [[2]]
#> T I A A L G N I I Y R A I E 
#> 6 3 5 5 3 1 6 3 3 4 2 5 3 6 
#> 
#> [[3]]
#> N Y E R T G H L I 
#> 6 4 6 2 6 1 5 3 3 
#> 
#> [[4]]
#> M A Y N N N I A L N 
#> 5 5 4 6 6 6 3 5 3 6 
#> 
#> [[5]]
#> M N 
#> 5 6 
#> 
#> [[6]]
#> N A A A T 
#> 6 5 5 5 6 
#> 

## Sequence with an extended alphabet, encoding with a standard alphabet
encsq_to_list(e5)

#> [[1]]
#>  V  A  H  C  H  A  G  D  U  G  B  B  V  G 
#> NA  1 NA  2 NA  1  2 NA NA  2 NA NA NA  2 
#> 
#> [[2]]
#>  C  U  G  C  V  B 
#>  2 NA  2  2 NA NA 
#> 
#> [[3]]
#>  D  U  U  B  G  A 
#> NA NA NA NA  2  1 
#> 
#> [[4]]
#>  C  C  C  U 
#>  2  2  2 NA 
#> 
#> [[5]]
#>  C  U  H  A  A  B  B  U 
#>  2 NA NA  1  1 NA NA NA 
#> 

encsq_to_list(e6)

#> [[1]]
#>  M  A  Y  U  O  U  O  N  N  N  I  A  L  N 
#>  5  5  4 NA NA NA NA  6  6  6  3  5  3  6 
#> 
#> [[2]]
#>  U  U  M  X  B  Z  O  N  O 
#> NA NA  5 NA NA NA NA  6 NA 
#> 
#> [[3]]
#> N A A G A T 
#> 6 5 5 1 5 6 
#> 

## Sequence with an extended alphabet, encoding with an extended alphabet
encsq_to_list(e7)

#> [[1]]
#>  V  A  H  C  H  A  G  D  U  G  B  B  V  G 
#> 21  1 13  1 13  1  2  8  3  2  5  5 21  2 
#> 
#> [[2]]
#>  C  U  G  C  V  B 
#>  1  3  2  1 21  5 
#> 
#> [[3]]
#> D U U B G A 
#> 8 3 3 5 2 1 
#> 
#> [[4]]
#> C C C U 
#> 1 1 1 3 
#> 
#> [[5]]
#>  C  U  H  A  A  B  B  U 
#>  1  3 13  1  1  5  5  3 
#> 

encsq_to_list(e8)

#> [[1]]
#>      M      A      Y      U      O      U      O      N      N      N      I 
#>  5.000  5.000  4.000 -0.100  0.753 -0.100  0.753  6.000  6.000  6.000  3.000 
#>      A      L      N 
#>  5.000  3.000  6.000 
#> 
#> [[2]]
#>       U       U       M       X       B       Z       O       N       O 
#>  -0.100  -0.100   5.000 -53.950   7.770   0.000   0.753   6.000   0.753 
#> 
#> [[3]]
#> N A A G A T 
#> 6 5 5 1 5 6 
#> 

## Sequence with standard alphabet, encoding without all letters assigned
encsq_to_list(e9)

#> [[1]]
#>    T    A    T    G    A    A    T    T    A    G    C    T    G    T    C    T 
#>   NA 1.00   NA 0.02 1.00 1.00   NA   NA 1.00 0.02   NA   NA 0.02   NA   NA   NA 
#>    T    T    G    C    T    G    C    T    T    T    G    G    T    T    A    T 
#>   NA   NA 0.02   NA   NA 0.02   NA   NA   NA   NA 0.02 0.02   NA   NA 1.00   NA 
#>    C    T    A    T    G    A 
#>   NA   NA 1.00   NA 0.02 1.00 
#> 
#> [[2]]
#>    C    T    T    T    G    G    T    T    A    T    C    T    A    G    C    T 
#>   NA   NA   NA   NA 0.02 0.02   NA   NA 1.00   NA   NA   NA 1.00 0.02   NA   NA 
#>    G    T    A    T    G    A 
#> 0.02   NA 1.00   NA 0.02 1.00 
#> 
#> [[3]]
#>    T    A    T    C    T    A    G    C    T    G    T    A    T    G 
#>   NA 1.00   NA   NA   NA 1.00 0.02   NA   NA 0.02   NA 1.00   NA 0.02 
#> 
#> [[4]]
#>    C    T    G    C    T    G 
#>   NA   NA 0.02   NA   NA 0.02 
#> 
#> [[5]]
#>    C    T    T    A    G    A 
#>   NA   NA   NA 1.00 0.02 1.00 
#> 
#> [[6]]
#>  C  C  C  T 
#> NA NA NA NA 
#> 
#> [[7]]
#>    C    T    G    A    A    T    G    T 
#>   NA   NA 0.02 1.00 1.00   NA 0.02   NA 
#> 

encsq_to_list(e10)

#> [[1]]
#>    N    Y    M    I    T    G    G    R    E    E    Y    E    R    T    V    I 
#>  6.0   NA  5.0  0.3   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA   NA  0.3 
#>    Y    R    A    I    A    L    N    A    A    N    Y    T    W    I    L 
#>   NA   NA  5.0  0.3  5.0 -3.1  6.0  5.0  5.0  6.0   NA   NA   NA  0.3 -3.1 
#> 
#> [[2]]
#>    T    I    A    A    L    G    N    I    I    Y    R    A    I    E 
#>   NA  0.3  5.0  5.0 -3.1   NA  6.0  0.3  0.3   NA   NA  5.0  0.3   NA 
#> 
#> [[3]]
#>    N    Y    E    R    T    G    H    L    I 
#>  6.0   NA   NA   NA   NA   NA  5.0 -3.1  0.3 
#> 
#> [[4]]
#>    M    A    Y    N    N    N    I    A    L    N 
#>  5.0  5.0   NA  6.0  6.0  6.0  0.3  5.0 -3.1  6.0 
#> 
#> [[5]]
#> M N 
#> 5 6 
#> 
#> [[6]]
#>  N  A  A  A  T 
#>  6  5  5  5 NA 
#> 

## Import encoding from other sources
encsq_to_list(e_aa)

#> [[1]]
#>     N     Y     M     I     T     G     G     R     E     E     Y     E     R 
#> 0.229 0.429 0.375 0.411 0.388 0.312 0.312 0.362 0.262 0.262 0.429 0.262 0.362 
#>     T     V     I     Y     R     A     I     A     L     N     A     A     N 
#> 0.388 0.495 0.411 0.429 0.362 0.288 0.411 0.288 0.400 0.229 0.288 0.288 0.229 
#>     Y     T     W     I     L 
#> 0.429 0.388 0.231 0.411 0.400 
#> 
#> [[2]]
#>     T     I     A     A     L     G     N     I     I     Y     R     A     I 
#> 0.388 0.411 0.288 0.288 0.400 0.312 0.229 0.411 0.411 0.429 0.362 0.288 0.411 
#>     E 
#> 0.262 
#> 
#> [[3]]
#>     N     Y     E     R     T     G     H     L     I 
#> 0.229 0.429 0.262 0.362 0.388 0.312 0.200 0.400 0.411 
#> 
#> [[4]]
#>     M     A     Y     N     N     N     I     A     L     N 
#> 0.375 0.288 0.429 0.229 0.229 0.229 0.411 0.288 0.400 0.229 
#> 
#> [[5]]
#>     M     N 
#> 0.375 0.229 
#> 
#> [[6]]
#>     N     A     A     A     T 
#> 0.229 0.288 0.288 0.288 0.388 
#> 

# What encoding is assigned to a letter

ee1 <- encsq_to_list(e1)
ee1[[1]]["T"]

#> T 
#> 2