Compare sq objects

Compares input sq object with another given.

# S3 method for sq
==(x1, x2)

Arguments

x1
an sq object.

x2
an object (a character vector or an sq object) to compare with x1.

Value

A logical vector indicating on which positions the objects are equal

Details

`==` converts left-hand side of the comparison (x1) to a character vector using as.character and checks whether object given on the right side can be compared with sq object. Function also checks the type of sq object with which given object is to be compared. If the type of sq object is "ami", "dna" or "rna" and given sequence is a character vector with any lowercase letters, the function replaces them with corresponding capital ones using toupper. If right-hand side object (x2) is of class sq, then it is converted to a character vector also using as.character function.

When both objects are already converted to character vectors, comparison is carried out element-wise with standard R rules, (e.g., recycling is used). You can check details at Comparison.

Comparing sequences as characters vectors cause that various types of sequences can be compared for example amino acid with nucleotide sequences so attention should be paid, which sequence types are compared.

See also

Examples

# Creating sq object to work on:
sq <- construct_sq(c("ACTGCTG", "CTTAGA", 
                     "CCCT", "CTGAATGT"), type = "dna")
                     
sq_different_len <- construct_sq(c("ACTGCTG", "CTTAGA", 
                                   "GGAA"), type = "dna")
                                   
sq_the_same_len <- construct_sq(c("ACTGCTG", "CTTAGA", 
                                 "CCCT", "CTGAATGT"), type = "dna")
                                                                                       
# Get an overview of the sequences:
summary(sq)

#>      Length Class  Mode
#> [1,] 3      -none- raw 
#> [2,] 3      -none- raw 
#> [3,] 2      -none- raw 
#> [4,] 3      -none- raw 

summary(sq_the_same_len)

#>      Length Class  Mode
#> [1,] 3      -none- raw 
#> [2,] 3      -none- raw 
#> [3,] 2      -none- raw 
#> [4,] 3      -none- raw 

summary(sq_different_len)

#>      Length Class  Mode
#> [1,] 3      -none- raw 
#> [2,] 3      -none- raw 
#> [3,] 2      -none- raw 

# Comparing sq object with an object of the same length :
sq == sq_the_same_len

#> [1] TRUE TRUE TRUE TRUE

# Comparing object sq object with an object of a different length : 
sq == sq_different_len

#> Warning: longer object length is not a multiple of shorter object length

#> [1]  TRUE  TRUE FALSE FALSE

 
# Comparing sq object to a given character vector of a different length:
sq == c('AAA','CCC')

#> [1] FALSE FALSE FALSE FALSE

# Comparing sq object to a given character vector of the same length:
sq == c("ACTGCTG", "CTTAGA",'CCCT', 'CTGAATGT')

#> [1] TRUE TRUE TRUE TRUE

# Comparing sq object to a given DNA element 'ATGTGA':
sq == 'ATGTGA'

#> [1] FALSE FALSE FALSE FALSE

# Comparing sq object to a given amino acid vector:
sq == c('RISGQQD','RISGQQD')

#> [1] FALSE FALSE FALSE FALSE