Export sq objects into other formats

Convert object of class sq to another class from another package. Currently supported packages are ape with its formats (AAbin and DNAbin), Bioconductor (AAStringSet, DNAStringSet) and seqinr (SeqFastaAA, SeqFastadna).

export_sq(sq, export_format, name)

Arguments

sq
a sq object.

export_format
a character string indicating package and the destination class; it should be one of the following: "seqinr::SeqFastaAA", "ape::AAbin", "Biostrings::AAStringSet", "seqinr::SeqFastadna", "ape::DNAbin", "Biostrings::DNAStringSet".

name
a character vector of length equal to sq length.

See also

Examples

sq_ami <- construct_sq(c("MVVGL", "LAVPP"))
export_sq(sq_ami, "ape::AAbin")

#> Error: you need to install 'ape' package to export object to its formats

export_sq(sq_ami, "Biostrings::AAStringSet", c("one", "two"))

#> Error: you need to install 'Biostrings' package to export object to its formats

export_sq(sq_ami, "seqinr::SeqFastaAA")

#> Error: you need to install 'seqinr' package to export object to its formats

sq_dna <- construct_sq(c("TGATGAAGCGCA", "TTGATGGGAA"))
export_sq(sq_dna, "ape::DNAbin", name = c("one", "two"))

#> Error: you need to install 'ape' package to export object to its formats

export_sq(sq_dna, "Biostrings::DNAStringSet")

#> Error: you need to install 'Biostrings' package to export object to its formats

export_sq(sq_dna, "seqinr::SeqFastadna")

#> Error: you need to install 'seqinr' package to export object to its formats