Test sq object for presence of given motifs

Test if elements of a sq object contain given motifs

x %has% y

Arguments

x
a sq object to be tested.

y
a character vector of motifs to be searched for.

Value

A logical vector of the same length as input sq, indicating which elements contain all given motifs

Details

This function allows testing if elements of a sq object contain a given motif or motifs, which includes start sequences and stop codons. It returns a logical for every element of the sq object - TRUE if it contains the motif and FALSE otherwise. When multiple motifs are searched, TRUE will be returned only for sequences that contain all the given motifs.

This function only indicates if a motif is present within a sequence, to find all motifs, and their positions within sequences use find_motifs.

Allowed and forbidden letters and characters details

Note: if a sq object contains characters: ^$?=()\.|+*[] in its alphabet, search for motifs cannot be performed and an error will be displayed (with exception of sq objects of type ami - in their alphabet there is '*' letter and it can be contained in sought motif). To search for motifs with those characters, you have to replace them first using substitute_letters.

In case of sq objects of type ami, dna and rna, motifs have to consist of upper case letters from amino acid, DNA and RNA alphabets respectively. Use of lower case letters will return an error. Two additional characters are allowed: '^' and '$' indicating the beginning and the end of a sequence respectively. Moreover, notice that '*' character may be used in amino acid motifs, as it is a part of the amino acid alphabet. If a motif contains ambiguous letters, all possible matches will be searched for, e.g., amino acid motif "MAJ" (where "J" is an ambiguous letter indicating L, or I) will find motifs: "MAJ", "MAL" and "MAI".

Detailed list of all letters corresponding to each ambiguous letter may be found at aminoacids_df and nucleotides_df.

See also

Examples

# Creating objects to work on:
sq_dna <- construct_sq(c("CTGAATGCAGTACCGTAAT", "ATGCCGTAAATGCCAT", 
                         "CAGACCANNNATAG"), type = 'dna')
sq_ami <- construct_sq(c("AGNTYIKFGGAYTI", "MATEGILIAADGYTWIL", 
                         "MIPADHICAANGIENAGIK"), type = 'ami')
                         
# Test if DNA sequences contain start codon 'ATG':
sq_dna %has% 'ATG'

#> [1]  TRUE  TRUE FALSE

# Test if DNA sequences begin with start codon 'ATG'
sq_dna %has% '^ATG'

#> [1] FALSE  TRUE FALSE

# Test if DNA sequences end with 'TAG' (one of the stop codons):
sq_dna %has% 'TAG$'

#> [1] FALSE FALSE  TRUE

# Test if amino acid sequences contain motif of two alanines followed by
# aspartic acid or asparagine ('AAB' motif will match 'AAB', 'AAD' and 'AAN'):
sq_ami %has% 'AAB'

#> [1] FALSE  TRUE  TRUE

# Test if amino acid sequences contain two motifs:
sq_ami %has% c("AAXG", "mat")

#> [1] FALSE  TRUE FALSE