Import sq objects from other objects

Creates sq object from object of class from another package. Currently supported packages are ape with its formats (AAbin and DNAbin), Bioconductor (AAStringSet, DNAStringSet) and seqinr (SeqFastaAA, SeqFastadna).

import_sq(object)

Arguments

object
- an object of one of classes: AAbin, DNAbin, AAStringSet, DNAStringSet, SeqFastaAA, SeqFastadna.

Value

A tibble with sq column of sq type representing the same sequences as given object; the object has a type corresponding to the input type; if given sequences had names, output tibble has also another column name with those names

Details

Providing object of class other than specified will result in error.

See also

Examples

## ape example
library(ape)

#> 
#> Attaching package: ‘ape’

#> The following object is masked from ‘package:tidysq’:
#> 
#>     complement

ape_dna <- as.DNAbin(list(one = c("C", "T", "C", "A"), two = c("T", "G", "A", "G", "G")))
import_sq(ape_dna)

#> # A tibble: 2 x 2
#>   name  sq       
#>   <chr> <(c)dna> 
#> 1 one   CTCA  <4>
#> 2 two   TGAGG <5>

## Biostrings example
library(Biostrings)

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: ‘BiocGenerics’

#> The following objects are masked from ‘package:parallel’:
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: ‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: ‘IRanges’

#> The following object is masked from ‘package:tidysq’:
#> 
#>     reverse

#> Loading required package: XVector

#> 
#> Attaching package: ‘Biostrings’

#> The following object is masked from ‘package:ape’:
#> 
#>     complement

#> The following object is masked from ‘package:tidysq’:
#> 
#>     complement

#> The following object is masked from ‘package:base’:
#> 
#>     strsplit

Biostrings_dna <- DNAStringSet(c(one = "CTCA", two = "TGAGG"))
import_sq(Biostrings_dna)

#> # A tibble: 2 x 2
#>   name  sq       
#>   <chr> <(c)dna> 
#> 1 one   CTCA  <4>
#> 2 two   TGAGG <5>

## seqinr example
library(seqinr)

#> 
#> Attaching package: ‘seqinr’

#> The following object is masked from ‘package:Biostrings’:
#> 
#>     translate

#> The following objects are masked from ‘package:ape’:
#> 
#>     as.alignment, consensus

seqinr_dna <- as.SeqFastadna(list(one = c("C", "T", "C", "A"), 
                                  two = c("T", "G", "A", "G", "G")))
import_sq(seqinr_dna)

#> # A tibble: 2 x 1
#>   sq       
#>   <(c)dna> 
#> 1 CTCA  <4>
#> 2 TGAGG <5>