Generate random sequences

Generates a sq object with specified number of sequences of given length and given type.

random_sq(n, len, type, is_clean, sd = NULL, use_gap = FALSE)

Arguments

n
a positive integer value - number of sequences to generate.

len
a positive integer value - length of each sequence if sd not specified and mean length of sequences if sd specified

type
a type of generated sq object; possible values are "ami" and "nuc" (see section sq types in sq documentation for details).

is_clean
a logical value - if TRUE, letters will be drawn from clean alphabet, if FALSE - ambiguous letters might be also generated.

sd
a positive numeric value; if specified, gives standard deviation of length of generated sequences.

use_gap
- a logical value; if TRUE, sequences will be generated with random gaps inside.

Value

An object of class sq with type as specified.

Sequences are generated using sample function. There is no possibility of generating a sequence of length 0, even if sd is given. Letter '*' is not used in generating ami sequences.

See also

Examples

# setting seed for reproducibility
set.seed(16)

# generating random sequences
random_sq(10, 10, "ami", TRUE)

#> ami (amino acids), cln (cleaned) sequences list:
#>  [1] ARMIYQYTLD                                                             <10>
#>  [2] GKWPIQYIMI                                                             <10>
#>  [3] QTMRSGWGMP                                                             <10>
#>  [4] HKVVAKKAFE                                                             <10>
#>  [5] FFKCDYQKMH                                                             <10>
#>  [6] SEFDRVYMDT                                                             <10>
#>  [7] FANDHVQKLE                                                             <10>
#>  [8] LCQAHYVPQW                                                             <10>
#>  [9] RLNAYHDNMT                                                             <10>
#> [10] CWKKVEESLW                                                             <10>
#> 

random_sq(50, 8, "nuc", FALSE, sd = 3)

#> nuc (nucleotides) sequences list:
#>  [1] MAYSSN                                                                  <6>
#>  [2] KCNTSCCT                                                                <8>
#>  [3] YNVRRRUVCHW                                                            <11>
#>  [4] HMUAYSHRG                                                               <9>
#>  [5] YDRKD                                                                   <5>
#>  [6] BMWAAMNBBGG                                                            <11>
#>  [7] M                                                                       <1>
#>  [8] KTDMBBSKNM                                                             <10>
#>  [9] KDKNGC                                                                  <6>
#> [10] GWTAMDWUAWBS                                                           <12>
#> printed 10 out of 50

random_sq(6, 100, "ami", TRUE, use_gap = TRUE)

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] MPYPSHT-RTSGPPIAGATYN-IASDRQDMDKNFYLCAFTWNIEGLQDVFQ-MCKWDRDHFECRFY...  <100>
#> [2] LVFGCFPSKYV-KLCYRQLKPLFNKNGMLSYENIREEINEYWVTKNTHMGHD-SCRPSFADILKSL...  <100>
#> [3] TKKENY-TCYGKNHDFMDEGDIAD-SQ-RERRYKASEMGPMWDPSEISFYAT-RIIANNH-SEYYE...  <100>
#> [4] WIWAHACTVMCWHYRDPQHAWH-RDRESWCKMLAYAYSAMSFKQSQYQDGSDDDWISTLTITPCIR...  <100>
#> [5] THAHIIDTVCRIAIHCFNWS-YDQEWPKIVIADS-EVKAEDVDRTH--HQCTSTERCQDMHH-VDP...  <100>
#> [6] METDSQIYMIIDPTIPMPMHFGAQNPDWLD-RWYCATPYGCKCAEIPEMKTL-KLLELVRLNLM-A...  <100>
#>