Generate random sequences

Generates an sq object with specified number of sequences of given length and alphabet.

random_sq(n, len, alphabet, sd = NULL, use_gap = FALSE)

Arguments

n
[integer(1)]
A number of sequences to generate - must be non-negative.

len
[integer(1)]
Length of each sequence if sd not specified and mean length of sequences if sd specified - must be non-negative.

alphabet
[character]
If provided value is a single string, it will be interpreted as type (see details). If provided value has length greater than one, it will be treated as atypical alphabet for sq object and sq type will be atp.

sd
[integer(1)]
If specified, gives standard deviation of length of generated sequences - must be non-negative.

use_gap
[logical(1)]
If TRUE, sequences will be generated with random gaps inside (commonly denoted as "-").

Value

An object of class sq with type as specified.

Details

Letter '*' is not used in generating ami sequences. If parameter sd is passed, then all generated negative values are replaced with 0s.

See also

Examples

# Setting seed for reproducibility
set.seed(16)

# Generating random sequences
random_sq(10, 10, "ami_bsc")

#> basic amino acid sequences list:

#>  [1] GSCSLRWKLV                                                             <10>
#>  [2] QFVGDSLHHW                                                             <10>
#>  [3] NKECSGNNID                                                             <10>
#>  [4] YLHLGCRLLQ                                                             <10>
#>  [5] YAAWCIMRTK                                                             <10>
#>  [6] MELQLGKIFK                                                             <10>
#>  [7] SNELCVTCVW                                                             <10>
#>  [8] FCHDEWFQAT                                                             <10>
#>  [9] WFNMFMQVRT                                                             <10>
#> [10] FKEGHSHCCN                                                             <10>

random_sq(25, 18, "rna_bsc", sd = 6)

#> basic RNA sequences list:

#>  [1] CCGCUCGCCGCGUGC                                                        <15>
#>  [2] CUGCCUUCUUGCUGCCCCUUC                                                  <21>
#>  [3] CCCUGCUCUCUCGCCGUUGUG                                                  <21>
#>  [4] UCCCCGCCGUUGUCU                                                        <15>
#>  [5] UGUUUUCCCCUGGG                                                         <14>
#>  [6] CUGCUUUGCGCGCCCGGGCGUGUGUGC                                            <27>
#>  [7] UCUUUUGGUUUGUUGCCUGC                                                   <20>
#>  [8] UUGCGUCGCGUGUGGGUCUU                                                   <20>
#>  [9] UCGUGUCGUCCUGUUCCGGG                                                   <20>
#> [10] CGUGCCGUGC                                                             <10>

#> printed 10 out of 25

random_sq(50, 8, "dna_ext", sd = 3)

#> extended DNA sequences list:

#>  [1] YAMAKGDD                                                                <8>
#>  [2] RACBTTGTYB                                                             <10>
#>  [3] YSTCCRC                                                                 <7>
#>  [4] RDBYNYYBBMA                                                            <11>
#>  [5] YCWVGVTCKYBKT                                                          <13>
#>  [6] DG                                                                      <2>
#>  [7] NWVBB                                                                   <5>
#>  [8] DSKNK                                                                   <5>
#>  [9] DBNMWRA                                                                 <7>
#> [10] GSKKYYNMT                                                               <9>

#> printed 10 out of 50

random_sq(6, 100, "ami_bsc", use_gap = TRUE)

#> basic amino acid sequences list:

#> [1] SRC-GSGTIILSSEIVGQYYMWHDNNMSMQIGQQRILFNMAMRAI-KHWDDGFLCYHGWMY-SYGYC... <100>
#> [2] IF-RRHYSFSLLEMQRLEILQWNWMIQMKTW-NQVTCANKGTIHDTGEHCFNAANTDCLA-HTLKHR... <100>
#> [3] VHRGHH-SSAIRQVDQENFFLSMQHNCQLESGDDKFC-QLMDWFEMMYETGSF-VFIDEWGQQFENG... <100>
#> [4] EHEADEGWCHVLASDRHYTGEEHMQMWCAWF-TAEFRNMLQWLGGWC-EHKLAKQWYET-SQNKIWD... <100>
#> [5] AC-G-WFHAVKFRCT-IRNVF---MCESSNAEFLHMVMWGHFHEELQMHFVWNTMHCHTYKFYRNKD... <100>
#> [6] MERQCLYHWGVTQNCHMKRQGSHDQW-EMLRARKSAHGEMCRSRMFFMKIDVWKGMNWGW-VDGLKR... <100>

# Passing whole alphabet instead of type
random_sq(4, 12, c("Pro", "Gly", "Ala", "Met", "Cys"))

#> atp (atypical alphabet) sequences list:

#> [1] Ala Gly Met Met Ala Gly Gly Cys Gly Ala Ala Ala                         <12>
#> [2] Ala Gly Cys Cys Gly Cys Met Cys Gly Gly Cys Met                         <12>
#> [3] Gly Met Cys Ala Ala Cys Cys Ala Met Gly Ala Ala                         <12>
#> [4] Met Met Met Ala Met Gly Ala Ala Met Cys Gly Ala                         <12>

# Generating empty sequences (why would anyone though)
random_sq(8, 0, "rna_ext")

#> extended RNA sequences list:

#> [1] <NULL>                                                                   <0>
#> [2] <NULL>                                                                   <0>
#> [3] <NULL>                                                                   <0>
#> [4] <NULL>                                                                   <0>
#> [5] <NULL>                                                                   <0>
#> [6] <NULL>                                                                   <0>
#> [7] <NULL>                                                                   <0>
#> [8] <NULL>                                                                   <0>