Generate random sequences

Generates a sq object with specified number of sequences of given length and given type.

random_sq(n, len, type, is_clean, sd = NULL, use_gap = FALSE)

Arguments

n
a positive integer value - number of sequences to generate.

len
a positive integer value - length of each sequence if sd not specified and mean length of sequences if sd specified

type
a type of generated sq object; possible values are "ami", "dna" and "rna" (see section sq types in sq documentation for details).

is_clean
a logical value - if TRUE, letters will be drawn from clean alphabet, if FALSE - ambiguous letters might be also generated.

sd
a positive numeric value; if specified, gives standard deviation of length of generated sequences.

use_gap
- a logical value; if TRUE, sequences will be generated with random gaps inside.

Value

An object of class sq with type as specified.

Sequences are generated using sample function. There is no possibility of generating a sequence of length 0, even if sd is given. Letter '*' is not used in generating ami sequences.

See also

Examples

# setting seed for reproducibility
set.seed(16)

# generating random sequences
random_sq(10, 10, "ami", TRUE)

#> ami (amino acids), cln (cleaned) sequences list:
#>  [1] ARMIYQYTLD                                                             <10>
#>  [2] GKWPIQYIMI                                                             <10>
#>  [3] QTMRSGWGMP                                                             <10>
#>  [4] HKVVAKKAFE                                                             <10>
#>  [5] FFKCDYQKMH                                                             <10>
#>  [6] SEFDRVYMDT                                                             <10>
#>  [7] FANDHVQKLE                                                             <10>
#>  [8] LCQAHYVPQW                                                             <10>
#>  [9] RLNAYHDNMT                                                             <10>
#> [10] CWKKVEESLW                                                             <10>

random_sq(25, 18, "rna", TRUE, sd = 6)

#> rna (RNA), cln (cleaned) sequences list:
#>  [1] UCCCACCUGUGAGU                                                         <14>
#>  [2] UUGUGGUAAGUGAUCAUG                                                     <18>
#>  [3] CGAUGACUCAAAUCGGUAUAGGG                                                <23>
#>  [4] UACUUGCCUGUGCCCAGCCCC                                                  <21>
#>  [5] UAAGGCCGGAC                                                            <11>
#>  [6] AAUUCAAUUUUGGUAUAUUUGAUU                                               <24>
#>  [7] AAA                                                                     <3>
#>  [8] UGCGCUAUACAACUGCACACA                                                  <21>
#>  [9] UAAAGAUAAAUCC                                                          <13>
#> [10] GUUUGGCCCUGCACCGUGUGAGGAC                                              <25>
#> printed 10 out of 25

random_sq(50, 8, "dna", FALSE, sd = 3)

#> dna (DNA) sequences list:
#>  [1] YBCHRGBGNGM                                                            <11>
#>  [2] NKWT                                                                    <4>
#>  [3] SCVNWYTDBNNNT                                                          <13>
#>  [4] NHTRHW                                                                  <6>
#>  [5] NTTMVBGA                                                                <8>
#>  [6] BBY                                                                     <3>
#>  [7] RTGBKT                                                                  <6>
#>  [8] BSCSWGVSWN                                                             <10>
#>  [9] VCMS                                                                    <4>
#> [10] AWSKSCYS                                                                <8>
#> printed 10 out of 50

random_sq(6, 100, "ami", TRUE, use_gap = TRUE)

#> ami (amino acids), cln (cleaned) sequences list:
#> [1] AWH-RDRESWCKMLAYAYSAMSFKQSQYQDGSDDDWISTLTITPCIRWMYQHIWT-CMHSCGDQSLP... <100>
#> [2] S-YDQEWPKIVIADS-EVKAEDVDRTH--HQCTSTERCQDMHH-VDPERFCHHNMVYGIEE-GKWRP... <100>
#> [3] HFGAQNPDWLD-RWYCATPYGCKCAEIPEMKTL-KLLELVRLNLM-AACQHTPWEIDRMNKMPSVRM... <100>
#> [4] AC-VKFLPKDMEYKCVKNRPDT-SSYDPNHSRLMVKNLK-MLLWDWL-LKP-GFFQHPKLEVELDVF... <100>
#> [5] -QQTACNYEKRMQFQRSKLCSPENKLCYVHWIVAHVERL-FTGFYIEFRGPQHWYSKYYDGWNPQSM... <100>
#> [6] ITSPLNMPMFNPHLTLSQDSCSDH--LNLPAKPNAGS--GRYR-NVASGPKPTVYKLIIIWEITKP-... <100>