Generate random sequences

Generates an sq object with specified number of sequences of given length and alphabet.

random_sq(n, len, alphabet, sd = NULL, use_gap = FALSE)

Arguments

n
[integer(1)]
A number of sequences to generate - must be non-negative.

len
[integer(1)]
Length of each sequence if sd not specified and mean length of sequences if sd specified - must be non-negative.

alphabet
[character]
If provided value is a single string, it will be interpreted as type (see details). If provided value has length greater than one, it will be treated as atypical alphabet for sq object and sq type will be atp.

sd
[integer(1)]
If specified, gives standard deviation of length of generated sequences - must be non-negative.

use_gap
[logical(1)]
If TRUE, sequences will be generated with random gaps inside (commonly denoted as "-").

Value

An object of class sq with type as specified.

Details

Letter '*' is not used in generating ami sequences. If parameter sd is passed, then all generated negative values are replaced with 0s.

See also

Examples

# Setting seed for reproducibility
set.seed(16)

# Generating random sequences
random_sq(10, 10, "ami_bsc")

#> basic amino acid sequences list:

#>  [1] MQHVMHDYMF                                                             <10>
#>  [2] CYALSQVIDQ                                                             <10>
#>  [3] KRANWMHVNL                                                             <10>
#>  [4] VKVAMFGNTD                                                             <10>
#>  [5] EFDMCIKVYL                                                             <10>
#>  [6] CKIAHASNDV                                                             <10>
#>  [7] KTMMFAWTRS                                                             <10>
#>  [8] RKLIRLYGEH                                                             <10>
#>  [9] WFYKNIVTGF                                                             <10>
#> [10] MPIKDMVEWP                                                             <10>

random_sq(25, 18, "rna_bsc", sd = 6)

#> basic RNA sequences list:

#>  [1] CCACUUCGUGCGUCUCUCGGGAUUC                                              <25>
#>  [2] GGUACACUCUGCACA                                                        <15>
#>  [3] UUUUCGCAAUUG                                                           <12>
#>  [4] AUGGCCGGGUAGAUCGAU                                                     <18>
#>  [5] UAUGAACGCGGCCUCAGUCCCAAG                                               <24>
#>  [6] UCGAAUAUAUGCAAUGUGAAC                                                  <21>
#>  [7] GCACUCUAGGUAAACACCACAGGCGCAAG                                          <29>
#>  [8] CGAUUGGCGUACUCGAGGC                                                    <19>
#>  [9] AUUAGCGAUUCCCA                                                         <14>
#> [10] GAACUCAUGGGUAUGUCUUAAGGGUGGC                                           <28>

#> printed 10 out of 25

random_sq(50, 8, "dna_ext", sd = 3)

#> extended DNA sequences list:

#>  [1] MGARRCTKGY                                                             <10>
#>  [2] MHHWBHGB                                                                <8>
#>  [3] SKWRACHGD                                                               <9>
#>  [4] YVKSKDK                                                                 <7>
#>  [5] YKTA                                                                    <4>
#>  [6] RVDCNGRBCWKANS                                                         <14>
#>  [7] BRRCHNSSA                                                               <9>
#>  [8] HVNNDVB                                                                 <7>
#>  [9] SBHBSH                                                                  <6>
#> [10] VKBGVDDMW                                                               <9>

#> printed 10 out of 50

random_sq(6, 100, "ami_bsc", use_gap = TRUE)

#> basic amino acid sequences list:

#> [1] IKK-WIGFPHIHRYYLACMNCHGCKLNTRYK-ID-KTGQVTPKE-LSARWRGHL-DHNHM-AMIMKD... <100>
#> [2] MWYFN-SCSTVDCDREKGTPGGYLNKP-NPDLYCRMAETEIQ-TKRGENKDCACVQTWQVYIPYWSN... <100>
#> [3] DKMNFRPFRHCVIKQSVIMGDGWMQGSW-WGDKARQFRNNMAYFYRITPFIQTKTCAQDAC-LTA-C... <100>
#> [4] NIGQDWMCQEDYKIMMWVRMI-NACWQMKTCFL-NTWKGQS-ANHRLKQ-EW-CAQFKWSMDLFDGK... <100>
#> [5] TASCFFQNH-EGWSEHALIVWNLMNN-DIGQLYGVEWDQKANRDVMWEFIIHHCCFSCPMPQHNMLQ... <100>
#> [6] ATA-YAIHRMMNIGV-LPCQGKIFLYQDKNQV-IWY-IITFDVRWTTNVGVHQMSYPEEIQKKQCMA... <100>

# Passing whole alphabet instead of type
random_sq(4, 12, c("Pro", "Gly", "Ala", "Met", "Cys"))

#> atp (atypical alphabet) sequences list:

#> [1] Ala Cys Pro Met Met Gly Gly Gly Gly Pro Gly Pro                         <12>
#> [2] Ala Gly Pro Cys Met Met Cys Met Met Gly Ala Met                         <12>
#> [3] Gly Gly Cys Ala Cys Gly Ala Pro Gly Met Cys Pro                         <12>
#> [4] Met Cys Pro Ala Pro Pro Met Gly Gly Gly Pro Met                         <12>

# Generating empty sequences (why would anyone though)
random_sq(8, 0, "rna_ext")

#> extended RNA sequences list:

#> [1] <NULL>                                                                   <0>
#> [2] <NULL>                                                                   <0>
#> [3] <NULL>                                                                   <0>
#> [4] <NULL>                                                                   <0>
#> [5] <NULL>                                                                   <0>
#> [6] <NULL>                                                                   <0>
#> [7] <NULL>                                                                   <0>
#> [8] <NULL>                                                                   <0>