Remove sequences containing NA values

Removes sequences containing ambiguous elements or removes NA values from sequences in a sq object.

remove_na(sq, only_elements = FALSE)

Arguments

sq
a sq object of type ami or nuc.

only_elements
logical value indicating if only ambiguous elements (i.e., matching more than one amino acid/nucleotide) of sequences should be removed. If FALSE (default) whole sequences containing ambiguous elements are removed.

Value

A sq object with the same type as input type. Sequences not containing any NA values are left untouched.

Details

This function allows removal of sequences containing NA values. By default, whole sequences containing ambiguous elements are removed and NULL (empty) sequences are introduced in their place. If only_elements = TRUE then only NA values are removed from sequences in sq object. NULL values (empty sequences) can be identified using is_null_sq function.

NA may be introduced as a result of using functions like substitute_letters or bite. They also appear in sequences if you are reading file using read_fasta or constructing sq object from character vector with construct_sq in fast mode and there are letters in file or in strings other than specified.

See also

Examples

# Creating objects to work on:
sq_ami <- construct_sq(c("MIAANYTWIL","TIAALGNIIYRAIE",
                         "NYERTGHLI", "MAYXXXIALN"), type = "ami")
sq_nuc <- construct_sq(c("ATGCAGGA", "GACCGAACGAN",
                         "TGACGAGCTTA", "ACTNNAGCN"), type = "nuc")

# Substituting some letters with NA
sq_ami_sub <- substitute_letters(sq_ami, c(E = NA, R = NA))
sq_nuc_sub <- substitute_letters(sq_nuc, c(N = NA))

# Biting sequences out of range
sq_bitten <- bite(sq_ami, 1:15)

#> Warning: some sequences are subsetted with index bigger than length - NA introduced

# Printing them
sq_ami_sub

#> atp (atypical alphabet) sequences list:
#> [1] MIAANYTWIL                                                              <10>
#> [2] TIAALGNIIY!AI!                                                          <14>
#> [3] NY!!TGHLI                                                                <9>
#> [4] MAYXXXIALN                                                              <10>
#> 

sq_nuc_sub

#> atp (atypical alphabet) sequences list:
#> [1] ATGCAGGA                                                                 <8>
#> [2] GACCGAACGA!                                                             <11>
#> [3] TGACGAGCTTA                                                             <11>
#> [4] ACT!!AGC!                                                                <9>
#> 

# Removing sequences containing NA
remove_na(sq_ami_sub)

#> atp (atypical alphabet) sequences list:
#> [1] MIAANYTWIL                                                              <10>
#> [2] <NULL>                                                                   <0>
#> [3] <NULL>                                                                   <0>
#> [4] MAYXXXIALN                                                              <10>
#> 

remove_na(sq_nuc_sub)

#> atp (atypical alphabet) sequences list:
#> [1] ATGCAGGA                                                                 <8>
#> [2] <NULL>                                                                   <0>
#> [3] TGACGAGCTTA                                                             <11>
#> [4] <NULL>                                                                   <0>
#> 

remove_na(sq_bitten)

#> ami (amino acids) sequences list:
#> [1] <NULL>                                                                   <0>
#> [2] <NULL>                                                                   <0>
#> [3] <NULL>                                                                   <0>
#> [4] <NULL>                                                                   <0>
#> 

# Removing only NA elements
remove_na(sq_ami_sub, only_elements = TRUE)

#> atp (atypical alphabet) sequences list:
#> [1] MIAANYTWIL                                                              <10>
#> [2] TIAALGNIIYAI                                                            <12>
#> [3] NYTGHLI                                                                  <7>
#> [4] MAYXXXIALN                                                              <10>
#> 

remove_na(sq_nuc_sub, TRUE)

#> atp (atypical alphabet) sequences list:
#> [1] ATGCAGGA                                                                 <8>
#> [2] GACCGAACGA                                                              <10>
#> [3] TGACGAGCTTA                                                             <11>
#> [4] ACTAGC                                                                   <6>
#> 

remove_na(sq_bitten, TRUE)

#> ami (amino acids) sequences list:
#> [1] MIAANYTWIL                                                              <10>
#> [2] TIAALGNIIYRAIE                                                          <14>
#> [3] NYERTGHLI                                                                <9>
#> [4] MAYXXXIALN                                                              <10>
#>