Reverse sequence

Reverse given list of sequences.

reverse(sq)

Arguments

sq
a sq object.

Value

A sq object of the same type as input object but each sequence is reversed.

Details

The reverse function reverses each sequence in supplied sq object (e.q. transforms "MIAANYTWIL" to "LIWTYNAAIM"). Empty sequences are left with no effect. This operation does not change the type of the input object nor its alphabet.

Since the function reverse returns a sq object, the print function is implicitly called.

See also

Examples

# Creating sq objects using construct_sq:
sq_ami <- construct_sq(c("MIAANYTWIL","TIAALGNIIYRAIE", 
                         "NYERTGHLI", "MAYXXXIALN"), type = "ami")
sq_dna <- construct_sq(c("ATGCAGGA", "GACCGAACGAN", ""), type = "dna")
sq_unt <- construct_sq(c("ATGCAGGA!", "TGACGAGCTTA", "", "TIAALGNIIYRAIE"))

# Reverse sequences:
reverse(sq_ami)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("amisq"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("reverse"), generic = structure("reverse", package = "IRanges"),     package = "IRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop("invalid call in method dispatch to 'reverse' (no default method)",         domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘reverse’ for signature ‘"amisq"’

reverse(sq_dna)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("dnasq"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("reverse"), generic = structure("reverse", package = "IRanges"),     package = "IRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop("invalid call in method dispatch to 'reverse' (no default method)",         domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘reverse’ for signature ‘"dnasq"’

reverse(sq_unt)

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("untsq"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("reverse"), generic = structure("reverse", package = "IRanges"),     package = "IRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop("invalid call in method dispatch to 'reverse' (no default method)",         domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘reverse’ for signature ‘"untsq"’

# Reverse cleaned sequences:
reverse(clean(sq_ami))

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("amisq"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("reverse"), generic = structure("reverse", package = "IRanges"),     package = "IRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop("invalid call in method dispatch to 'reverse' (no default method)",         domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘reverse’ for signature ‘"amisq"’

reverse(clean(sq_dna))

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("dnasq"), new("standardGeneric", .Data = function (x,     ...) standardGeneric("reverse"), generic = structure("reverse", package = "IRanges"),     package = "IRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         ...)     stop("invalid call in method dispatch to 'reverse' (no default method)",         domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘reverse’ for signature ‘"dnasq"’