Apply function to each sequence

Applies given function to each sequence. Sequences are passed to function as character vectors (or numeric, if type of sq is enc) or single character strings, depending on parameter.

sqapply(sq, fun, ..., paste_char = FALSE, use_na_letter = paste_char)

Arguments

sq
a sq object.

fun
a function to apply to each sequence in sq object; it should take a character vector, numeric vector or single character string as an input.

...
another arguments passed to fun.

paste_char
a logical value indicating in which form sequences should be passed to the function fun; if FALSE (default), they will be treated as character vectors, if TRUE, they will be pasted into a single string.

use_na_letter
a logical value indicating whether to use a printing character to represent NA values; if TRUE, letter from option "tidysq_p_na_letter" will be used instead of NA values (default value for this option is "!", for details see tidysq-options), otherwise just NA values will be used; default value for this parameter is equal to paste_char value; use_na_letter cannot be FALSE if paste_char is TRUE.

Value

A list of values returned by function for each sequence in corresponding order.

See also

Examples

sq_ami <- construct_sq(c("YCYWIFTSRIK", "GGWGDVKCG", "KTKHIEQKL"))

# count how many times "K" appears in each sequence:
sqapply(sq_ami, function(sequence) sum(sequence == "K"))

#> [[1]]
#> [1] 1
#> 
#> [[2]]
#> [1] 1
#> 
#> [[3]]
#> [1] 3
#> 

# duplicate each element of sequence, then paste it and construct new sq
duplicated <- sqapply(sq_ami, function(sequence) paste(rep(sequence, each = 2), collapse = ""))
construct_sq(unlist(duplicated))

#> ami (amino acids), cln (cleaned) sequences list:

#> [1] YYCCYYWWIIFFTTSSRRIIKK                                                  <22>
#> [2] GGGGWWGGDDVVKKCCGG                                                      <18>
#> [3] KKTTKKHHIIEEQQKKLL                                                      <18>