Concatenate sq objects

Merges multiple sq and maybe character objects into one larger sq object.

Arguments

...
multiple sq and character objects. For exact behaviour, check Details section. First argument must be of sq class due to R mechanism of single dispatch. If this is a problem, recommended alternative is vec_c method from vctrs package.

Value

A sq object with length equal to sum of lengths of individual objects passed as parameters. Elements of sq are concatenated just as if they were normal lists (see c)

Details

Whenever all passed objects are either dnasq, rnasq or amisq, returned object is also of the same class. If all of them contain clnsq subtype, resulting object contains it as well, else clnsq subtype is dropped.

Mixing dnasq, rnasq and amisq is prohibited, as interpretation of symbols differ depending on the type.

Whenever all objects are either untsq or atpsq, returned object is also of the same class. These classes are not used with clnsq subtype, so it never appears in this context. Resulting object has alphabet equal to the set union of all alphabets of the involved objects.

Moreover, untsq objects may be concatenated with dnasq, rnasq and amisq objects, resulting in an untsq object with alphabet equal to the set union of all alphabets involved. However, user is strongly encouraged to use this possibility with caution, as it may result in unwanted concatenation of DNA and amino acid sequences.

Whenever character vectors are passed as second and later argument, they don't influence resulting sq object type. Each element of vector is interpreted as separate sequence. If resulting sq is predicted to have clnsq subtype, then passing character vector with characters not included in the resulting alphabet will cause code to fail with an exception. If result doesn't have clnsq subtype, all "foreign" characters will be silently replaced will NA.

Due to R dispatch mechanism passing character vector as first will return class-less list. This behaviour is effectively impossible and definitely unrecommended to fix, as fixing it would involve changing c primitive. If such possibility is necessary, vec_c is a better alternative.

Examples

cdnasq_1 <- construct_sq(c("GGACTGCA", "CTAGTA", ""), type = "dna")
cdnasq_2 <- construct_sq(c("ATGACA", "AC-G", "-CCAT"), type = "dna")
cdnasq_3 <- construct_sq(character(), type = "dna")
dnasq_1 <- construct_sq(c("BNACV", "GDBADHH"), type = "dna")
crnasq_1 <- construct_sq(c("UAUGCA", "UAGCCG"), type = "rna")
rnasq_1 <- construct_sq(c("-AHVRYA", "G-U-HYR"), type = "rna")
rnasq_2 <- construct_sq("AUHUCHYRBNN--", type = "rna")
camisq_1 <- construct_sq("ACHNK-IFK-VYW", type = "ami")
untsq_1 <- construct_sq("AF:gf;PPQ^&XN")

# Concatenating same-type sequences
# Only clean sequences
c(cdnasq_1, cdnasq_2, cdnasq_3)

#> dna (DNA), cln (cleaned) sequences list:

#> [1] GGACTGCA                                                                 <8>
#> [2] CTAGTA                                                                   <6>
#> [3] <NULL>                                                                   <0>
#> [4] ATGACA                                                                   <6>
#> [5] AC-G                                                                     <4>
#> [6] -CCAT                                                                    <5>

# Only not clean sequences
c(rnasq_1, rnasq_2)

#> rna (RNA) sequences list:

#> [1] -AHVRYA                                                                  <7>
#> [2] G-U-HYR                                                                  <7>
#> [3] AUHUCHYRBNN--                                                           <13>

# Both clean and unclean sequences
c(cdnasq_3, dnasq_1, cdnasq_2)

#> dna (DNA) sequences list:

#> [1] BNACV                                                                    <5>
#> [2] GDBADHH                                                                  <7>
#> [3] ATGACA                                                                   <6>
#> [4] AC-G                                                                     <4>
#> [5] -CCAT                                                                    <5>

# Mixing DNA and RNA sequences don't work
if (FALSE) {
c(cdnasq_1, crnasq_1)
}

# untsq can be mixed with DNA, RNA and amino acids
c(camisq_1, untsq_1)

#> unt (unspecified type) sequences list:

#> [1] ACHNK-IFK-VYW                                                           <13>
#> [2] AF:gf;PPQ^&XN                                                           <13>

c(untsq_1, crnasq_1, rnasq_1)

#> unt (unspecified type) sequences list:

#> [1] AF:gf;PPQ^&XN                                                           <13>
#> [2] UAUGCA                                                                   <6>
#> [3] UAGCCG                                                                   <6>
#> [4] -AHVRYA                                                                  <7>
#> [5] G-U-HYR                                                                  <7>

c(cdnasq_2, untsq_1, cdnasq_3)

#> unt (unspecified type) sequences list:

#> [1] ATGACA                                                                   <6>
#> [2] AC-G                                                                     <4>
#> [3] -CCAT                                                                    <5>
#> [4] AF:gf;PPQ^&XN                                                           <13>

# Character vectors are also acceptable
c(cdnasq_2, "TGCA-GA")

#> dna (DNA), cln (cleaned) sequences list:

#> [1] ATGACA                                                                   <6>
#> [2] AC-G                                                                     <4>
#> [3] -CCAT                                                                    <5>
#> [4] TGCA-GA                                                                  <7>

c(rnasq_1, c("UACUGGGACUG", "AUGUBNAABNRYYRAU"), rnasq_2)

#> rna (RNA) sequences list:

#> [1] -AHVRYA                                                                  <7>
#> [2] G-U-HYR                                                                  <7>
#> [3] UACUGGGACUG                                                             <11>
#> [4] AUGUBNAABNRYYRAU                                                        <16>
#> [5] AUHUCHYRBNN--                                                           <13>

c(untsq_1, "&#JIA$O02t30,9ec", camisq_1)

#> unt (unspecified type) sequences list:

#> [1] AF:gf;PPQ^&XN                                                           <13>
#> [2] &!!IA!!!!!!!!!!!                                                        <16>
#> [3] ACHNK-IFK-VYW                                                           <13>