Set options of package

You can get value of an option by calling getOptions(option_name) and set its value by calling options(option_name = value), where option_name is an option name (full list of this package included below) and value is a value to assign to an option.

Details

You can change default behavior of package using one of following options:

tidysq_a_bite_na (default "warning") - a character string specifying in which way to inform user about biting sequences out of the range when using bite; possible values: "error", "warning", "message", "none",

tidysq_a_cln_sub_letters (default "warning") - a character string specifying in which way to inform user about dropping cln subtype of sq while using substitute_letters; possible values: "error", "warning", "message", "none",

tidysq_a_typify_small_cap_let (default "warning") - a character string specifying in which way to inform user about merging lowercase and uppercase letters when typifying sq object; possible values: "error", "warning", "message", "none",

tidysq_a_no_given_enc (default "warning") - a character string specifying in which way to inform user about encoding unspecified letters as NA if they do appear in sequences in encode; possible values: "error", "warning", "message", "none",

tidysq_g_fast_mode (default FALSE) - a logical value if to work in fast mode (see fast-mode)

tidysq_p_max_pillar_width (default 15) - an integer value specifying pillar_shaft_sq width

tidysq_p_max_sequences (default 10) - an integer value specifying maximum number of printed sequences in print sq, (see print sq)

tidysq_p_na_char (default = "!") - a character string which character string to pring when NA values appear in sequences,

tidysq_p_use_color (default TRUE) - a logical value if to use colorful printing

If value is not appropriate for given option, default value will be used.

See also